Return the available mutation CCF estimates in the data.

This function extracts Cancer Cell Fractions (CCFs) estimates from a `CNAqc` object, after they have been computed using function `compute_CCF`. The estimates are pooled across the used karyotypes, and extracted from the `CCF_estimates` field of the input object.

CCF(x)

Arguments

x
An object of class cnaqc, where CCF have been computed using function `compute_CCF`.

Value

A tibble with the mutations with the new columns for mutation multiplicity and CCF values.

Examples

data('example_dataset_CNAqc')
x = init(example_dataset_CNAqc$snvs, example_dataset_CNAqc$cna, example_dataset_CNAqc$purity)

#>  [ CNAqc - CNA Quality Check ] 
#> 

#> [CNAqc] Missing CCF column from CNA calls, adding CCF = 1 assuming all calls to be clonal.

#> Input  n = 13141 mutations for 267 CNA segments (267 clonal, 0 subclonal) 
#> Mapped. n = 12963 mutations mapped to clonal segments (~99% of input) 

x = compute_CCF(x)

#> 
#> =-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
#>  Computing mutation multiplicity with entropy H(x)
#> =-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
#> Binomial peaks  
#> # A tibble: 2 x 4
#>   mutation_multiplicity karyotype  peak label     
#>                   <dbl> <chr>     <dbl> <chr>     
#> 1                     1 2:1       0.308 One copy  
#> 2                     2 2:1       0.616 Two copies
#> H(x) quantile 0.9 
#>  Median depth 111 
#> Mixing proportions ([1, 99] quantiles):  0.554109589041096 0.445890410958904 
#> H(x)-derived cutoffs 0.441441441441441 0.468468468468468 
#> 

#> Warning: NAs introduced by coercion

#> Warning: Removed 2 rows containing non-finite values (stat_bin).

#> Warning: Removed 6 rows containing missing values (geom_bar).

#> Warning: Removed 2 rows containing missing values (geom_bar).

#> 
#> =-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
#>  Computing mutation multiplicity with entropy H(x)
#> =-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
#> Binomial peaks  
#> # A tibble: 2 x 4
#>   mutation_multiplicity karyotype  peak label     
#>                   <dbl> <chr>     <dbl> <chr>     
#> 1                     1 2:0       0.445 One copy  
#> 2                     2 2:0       0.89  Two copies
#> H(x) quantile 0.9 
#>  Median depth 76 
#> Mixing proportions ([1, 99] quantiles):  0.0933333333333333 0.906666666666667 
#> H(x)-derived cutoffs 0.657894736842105 0.710526315789474 
#> 

#> Warning: Removed 4 rows containing missing values (geom_bar).

#> Warning: Removed 2 rows containing missing values (geom_bar).

#> 
#> =-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
#>  Computing mutation multiplicity with entropy H(x)
#> =-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
#> Binomial peaks  
#> # A tibble: 2 x 4
#>   mutation_multiplicity karyotype  peak label     
#>                   <dbl> <chr>     <dbl> <chr>     
#> 1                     1 2:2       0.235 One copy  
#> 2                     2 2:2       0.471 Two copies
#> H(x) quantile 0.9 
#>  Median depth 141 
#> Mixing proportions ([1, 99] quantiles):  0.0946892325649256 0.905310767435074 
#> H(x)-derived cutoffs 0.319148936170213 0.347517730496454 
#> 

#> Warning: NAs introduced by coercion

#> Warning: Removed 38 rows containing non-finite values (stat_bin).

#> Warning: Removed 6 rows containing missing values (geom_bar).

#> Warning: Removed 2 rows containing missing values (geom_bar).

#> 
#> =-=-=-=-=-=-=-=-=-=-=-=-=-
#>  Summary CCF assignments
#> =-=-=-=-=-=-=-=-=-=-=-=-=-
#> # A tibble: 8 x 3
#>   karyotype mutation_multiplicity assignments
#>   <chr>                     <dbl>       <int>
#> 1 2:0                           1           9
#> 2 2:0                           2          72
#> 3 2:1                           1         886
#> 4 2:1                           2         675
#> 5 2:1                          NA           2
#> 6 2:2                           1        1013
#> 7 2:2                           2        6427
#> 8 2:2                          NA          38
#> Note: NA ~ mutations not confidently assignable at q = 0.9 

CCF(x)

#> # A tibble: 9,122 x 15
#>    chr     from     to ref   alt   FILTER    DP    NV    VAF ANNOVAR_FUNCTION
#>    <chr>  <dbl>  <dbl> <chr> <chr> <chr>  <dbl> <dbl>  <dbl> <chr>           
#>  1 chr2  3.58e5 3.58e5 C     A     PASS     104    58 0.558  intergenic      
#>  2 chr2  9.09e5 9.09e5 A     G     PASS      15     7 0.467  ncRNA_intronic  
#>  3 chr2  1.04e6 1.04e6 C     T     PASS      93    57 0.613  intronic        
#>  4 chr2  1.33e6 1.33e6 A     T     PASS     104    55 0.529  intronic        
#>  5 chr2  1.52e6 1.52e6 C     T     PASS      90    50 0.556  intronic        
#>  6 chr2  2.20e6 2.20e6 G     T     PASS     125    42 0.336  intronic        
#>  7 chr2  2.90e6 2.90e6 C     T     PASS     109    58 0.532  downstream      
#>  8 chr2  3.13e6 3.13e6 A     G     PASS     134    35 0.261  ncRNA_intronic  
#>  9 chr2  3.83e6 3.83e6 A     T     PASS     120    68 0.567  intergenic      
#> 10 chr2  3.88e6 3.88e6 T     A     PASS     126    10 0.0794 intergenic      
#> # … with 9,112 more rows, and 5 more variables: GENE <chr>, karyotype <chr>,
#> #   segment_id <chr>, mutation_multiplicity <dbl>, CCF <dbl>