CNAqc-based purity-optimisation pipeline for Sequenza CNA calling.

This pipeline implements optimised allele-specific copy number calling with Sequenza and CNAqc, optimising purity estimated from data, and allele-specific segments.

Requirements:

1. (Prerequisites) Run Sequenza utils steps outside R a. Process a FASTA file to produce a GC Wiggle track file b. Process BAM and Wiggle files to produce a seqz file: c. Post-process by binning the original seqz file:

Instructions to customize these steps are available at https://cran.r-project.org/web/packages/sequenza/vignettes/sequenza.html

2. Then this pipeline (Sequenza_CNAqc) will optimize CNA calling and purity estiamtion using peak detection from CNAqc. A maximum number of attempts max_runs is attempted, starting from a broad set of initial conditions for purity (5 range based on the score computed by analyse_peaks, which determines a purity gradient. Each run is saved in a folder named run-1, run-2, etc. If at any step a PASS is obtained via analyse_peaks, the pipeline stops and a symbolic link (final) pointing to the correct Sequenza solution is created. If no run reaches a PASS status, the final folder is never created but the best reported solution is print out to screen. Every CNAqc output object containing somatic mutations called by Sequenza and allele-specific CNAs is available in the specific running folders.

Sequenza_CNAqc(
  sample_id,
  seqz_file,
  dataframe = NULL,
  sex,
  cellularity = c(0.05, 1),
  ploidy = c(1.8, 5.4),
  max_runs = 3,
  reference = "GRCh38",
  normalization.method = "median",
  window = 1e+05,
  gamma = 280,
  kmin = 300,
  min.reads.baf = 50,
  min.reads = 50,
  min.reads.normal = 15,
  max.mut.types = 1,
  delta_cellularity = 0.05,
  delta_ploidy = 0.25,
  n_grid_cells = 10,
  verbose = FALSE
)

Arguments

sample_id

The id of the sample.

...

Optional parameters passed to the analyse_peaks function by CNAqc. Tune these to change error tolerance or karyotypes to use for QC.

Value

Nothing. If the pipeline could fit a correct purity value then a final folder links to the final Sequenza results. Otherwise the best solution - not final - is reported to screen. If the solution in general can not be found error logs and screen outputs suggest the user how to check, manually, how to improve the fits.

Note

Ploidy is kept constant by this pipeline, using a possible range of ploidy values that are specified by the user. If at any time a purity proposal made by CNAqc leads to inconsistent values (i.e., outside $[0,1]$) the pipeline stops prompting to the user to adjust manually the range of ploidy.

Examples