Analyze calls by peak detection.

This function carries out a peak-detection analysis to determine QC for clonal simple CNAs, clonal complex CNAs and subclonal CNAs.

Several scores are produced and returned; functions to visualize results are also available (see the tool website for example usages).

analyze_peaks(
  x,
  karyotypes = c("1:0", "1:1", "2:0", "2:1", "2:2"),
  min_karyotype_size = 0,
  min_absolute_karyotype_mutations = 100,
  p_binsize_peaks = 0.005,
  matching_epsilon = NULL,
  purity_error = 0.05,
  VAF_tolerance = 0.015,
  n_bootstrap = 1,
  kernel_adjust = 1,
  matching_strategy = "closest",
  KDE = TRUE,
  starting_state_subclonal_evolution = "1:1"
)

Arguments

x

An object of class cnaqc, created by the init function.

karyotypes

The list of karyotypes to test. By default LOH regions (A, AA), diploid regions (AB), and amplification regions (AAB, AABB). These correspond to '1:0', '1:1', '2:1', '2:0', '2:2' in "Major:minor" notation.

min_karyotype_size

Karyotypes are subset by their size (normalized for the number of input mutations). Karyotypes smaller than this value are removed from analysis.

min_absolute_karyotype_mutations

As min_karyotype_size this imposes a minimum number of mutations on a karyotype to be analysed (in absolute number of mutations). Karyotypes smaller than this value are removed from analysis.

p_binsize_peaks

Peaks detected will be filtered if, in a peak, we map less than p_binsize_peaks * N mutations. The value N is obtained couting all mutations that map in all peaks.#'

matching_epsilon

Deprecated parameter.

purity_error

Purity error tolerance to determine QC results. This can be set automatically using function auto_tolerance

VAF_tolerance

A tolerance in comparing bands overlaps - this is applied to the raw VAF values.

n_bootstrap

Number of times peak detection is bootstrapped (this helps sometimes finding peaks that might be visually observable but fail to be detected by the underlying peak detection heuristic). The default of this parameter is 1, meaning that no bootstrap is performed.

kernel_adjust

KDE adjust density parameter; see density. A Gaussian kernel is used (kernel = 'gaussian').

matching_strategy

If "closest", the closest peak will be used to match the expected peak. If "rightmost", peaks are matched prioritizing right to left peaks (the higher get matched first); this strategy is more correct in principle but works only if there are no spurious peaks in the estimated density. By defaule the first strategy is used.

KDE

Deprecated parameter (unused).

starting_state_subclonal_evolution

Starting state to determine linear versus branching evolutionary model, for sublonal CNA QC.

Value

An object of class cnaqc, modified to hold the results from this analysis. See the vignette to see how to extract and plot the results.

Examples

x = analyze_peaks(x)
#> 
#> ℹ Predicting coding
#> ── Peak analysis: simple CNAs ──────────────────────────────────────────────────
#> ℹ Predicting coding
#> 
#> ℹ Analysing 120065 mutations mapping to karyotype(s) 2:1, 2:2, 2:0, 1:1, and 1:0.
#> ℹ Predicting coding
#> ℹ Mixed type peak detection for karyotype 1:0 (124 mutations)
#> ℹ Predicting coding
#> ℹ Mixed type peak detection for karyotype 1:1 (855 mutations)
#> ℹ Predicting coding
#> ℹ Mixed type peak detection for karyotype 2:0 (5374 mutations)
#> ℹ Predicting coding
#> ℹ Mixed type peak detection for karyotype 2:1 (88422 mutations)
#> ℹ Predicting coding
#> ℹ Mixed type peak detection for karyotype 2:2 (25290 mutations)
#> ℹ Predicting coding
#> # A tibble: 8 × 16
#> # Rowwise: 
#>   mutation_multiplicity karyotype  peak delta_vaf     x     y counts_per_bin
#>                   <dbl> <chr>     <dbl>     <dbl> <dbl> <dbl>          <int>
#> 1                     1 2:1       0.268   0.0134  0.264 4.14            3138
#> 2                     2 2:1       0.537   0.0268  0.54  3               2729
#> 3                     1 2:2       0.212   0.00831 0.212 0.929            198
#> 4                     2 2:2       0.423   0.0166  0.423 6.85            1732
#> 5                     1 2:0       0.367   0.025   0.357 0.868             50
#> 6                     2 2:0       0.734   0.05    0.735 5.43             265
#> 7                     1 1:1       0.367   0.025   0.366 2.42              15
#> 8                     1 1:0       0.580   0.0624  0.570 4.16               7
#> # … with 9 more variables: discarded <lgl>, from <chr>, offset_VAF <dbl>,
#> #   offset <dbl>, weight <dbl>, epsilon <dbl>, VAF_tolerance <dbl>,
#> #   matched <lgl>, QC <chr>
#> ✔ Peak detection PASS with r = 0.00914757882473995 - maximum purity error ε = 0.05.
#> ℹ Predicting coding
#> 
#> ℹ Predicting coding
#> ── Peak analysis: complex CNAs ─────────────────────────────────────────────────
#> ℹ Predicting coding
#> 
#> ℹ Karyotypes 3:2, 3:1, 3:0, 3:3, 4:0, 4:2, 4:3, 5:2, 5:3, and 6:3 with >100 mutation(s). Using epsilon = 0.05.
#> ℹ Predicting coding
#> Error in arrange(., prop %>% desc): Problem with the implicit `transmute()` step.
#> ✖ Problem while computing `..1 = prop %>% desc`.
#> Caused by error:
#> ! unable to find an inherited method for function ‘desc’ for signature ‘"numeric"’
print(x)
#> ── [ CNAqc ]  293736 mutations in 667 segments (654 clonal, 13 subclonal). Genom
#> ℹ Predicting coding
#> 
#> ℹ Predicting coding
#> ── Clonal CNAs 
#> ℹ Predicting coding
#> 
#>   2:1  [n = 88422, L =   692 Mb] ■■■■■■■■■■■■■■■■■■■■■■■■■■■
#>   3:2  [n = 58384, L =   417 Mb] ■■■■■■■■■■■■■■■■■■  { BRAF }
#>   3:1  [n = 48704, L =   380 Mb] ■■■■■■■■■■■■■■■
#>   3:0  [n = 26622, L =   360 Mb] ■■■■■■■■  { CDKN2A }
#>   2:2  [n = 25290, L =   253 Mb] ■■■■■■■■
#>   3:3  [n = 16790, L =   115 Mb] ■■■■■
#>   2:0  [n =  5374, L =    67 Mb] ■■
#>   4:0  [n =  1752, L =    22 Mb] ■  { TP53 }
#>   4:2  [n =  1441, L =    11 Mb] 
#>   1:1  [n =   855, L =     9 Mb] 
#> 
#> 
#> ℹ Predicting coding
#> ── Subclonal CNAs (showing up to 10 segments) 
#> ℹ Predicting coding
#> 
#>  chr11@55700000  [n =  10468, L =  78.75 Mb]   2:1 (0.21)   2:2 (0.79) ■■■■■■■■■■
#>  chr11@17365005  [n =   5389, L =  31.55 Mb]   2:1 (0.21)   2:2 (0.79) ■■■■■
#>   chr11@5372292  [n =   1014, L =  11.99 Mb]   2:1 (0.21)   2:2 (0.79) 
#>    chr11@202253  [n =    610, L =   5.17 Mb]   2:1 (0.22)   2:2 (0.78) 
#>  chr11@48918601  [n =    542, L =   2.68 Mb]   2:1 (0.25)   2:2 (0.75) 
#>   chr6@82432583  [n =    301, L =   1.81 Mb]   2:1 (0.19)   2:2 (0.81) 
#>  chr11@51600000  [n =    290, L =    4.1 Mb]   2:1 (0.26)   2:2 (0.74) 
#>   chr6@81896364  [n =     69, L =   0.54 Mb]   2:1 (0.19)   2:2 (0.81) 
#>   chr6@93956180  [n =     41, L =   0.11 Mb]    2:1 (0.2)    2:2 (0.8) 
#>   chr8@42633277  [n =     13, L =   0.26 Mb]   2:1 (0.28)   2:2 (0.72) 
#> 
#> ℹ Sample Purity: 73.4% ~ Ploidy: 3.
#> ℹ Predicting coding
#> 
#> ℹ There are 3 annotated driver(s) mapped to clonal CNAs.
#> ℹ Predicting coding
#>          chr      from        to ref alt DP NV       VAF driver_label is_driver
#>        chr17   7577082   7577082   C   T 78 70 0.8974359         TP53      TRUE
#>         chr7 140453136 140453136   A   T 95 54 0.5684211         BRAF      TRUE
#>         chr9  21971120  21971120   G   A 23 14 0.6086957       CDKN2A      TRUE
#> 
#> ℹ Predicting coding
#> ──  PASS  Peaks QC closest: 100%, λ = -0.0104. Purity correction: -1%. ─────────
#> ℹ Predicting coding
#> 
#> ℹ 1:1 ~ n = 855   (  1%) →  PASS  -0.006 
#> ℹ Predicting coding
#> ℹ 2:0 ~ n = 5374  (  4%) →  PASS  -0.003   PASS  -0.002 
#> ℹ Predicting coding
#> ℹ 2:1 ~ n = 88422 ( 74%) →  PASS  -0.006   PASS  -0.006 
#> ℹ Predicting coding
#> ℹ 2:2 ~ n = 25290 ( 21%) →  PASS  -0.053   PASS  -0.004 
#> ℹ Predicting coding
#> 
#> ℹ Predicting coding
#> ── General peak QC (154052 mutations):  PASS  21  FAIL  3 - epsilon = 0.03. ────
#> ℹ Predicting coding
#> 
#> ℹ 3:0 ~ n = 26622 ( 17%) →  PASS  3  FAIL  0
#> ℹ Predicting coding
#> ℹ 3:1 ~ n = 48704 ( 32%) →  PASS  3  FAIL  0
#> ℹ Predicting coding
#> ℹ 3:2 ~ n = 58384 ( 38%) →  PASS  3  FAIL  0
#> ℹ Predicting coding
#> ℹ 3:3 ~ n = 16790 ( 11%) →  PASS  3  FAIL  0
#> ℹ Predicting coding
#> ℹ 4:0 ~ n = 1752  (  1%) →  PASS  3  FAIL  1
#> ℹ Predicting coding
#> ℹ 4:2 ~ n = 1441  (  1%) →  PASS  3  FAIL  1
#> ℹ Predicting coding
#> ℹ 4:3 ~ n = 359   (  0%) →  PASS  3  FAIL  1
#> ℹ Predicting coding
#> Error in arrange(., BR %>% desc, LI %>% desc): Problem with the implicit `transmute()` step.
#> ✖ Problem while computing `..1 = BR %>% desc`.
#> Caused by error:
#> ! unable to find an inherited method for function ‘desc’ for signature ‘"integer"’

print(x$peaks_analysis)
#> $score
#> [1] -0.01039476
#> 
#> $fits
#> $fits$`1:0`
#> $fits$`1:0`$matching
#> NULL
#> 
#> 
#> $fits$`1:1`
#> $fits$`1:1`$matching
#> # A tibble: 1 × 14
#> # Rowwise: 
#>   mutation_multiplic… karyotype  peak     x     y counts_per_bin discarded from 
#>                 <dbl> <chr>     <dbl> <dbl> <dbl>          <int> <lgl>     <chr>
#> 1                   1 1:1       0.367  0.37  2.56             21 FALSE     KDE  
#> # … with 6 more variables: offset_VAF <dbl>, offset <dbl>, weight <dbl>,
#> #   epsilon <dbl>, VAF_tolerance <dbl>, matched <lgl>
#> 
#> $fits$`1:1`$density
#> 
#> Call:
#> 	density.default(x = y, adjust = kernel_adjust, kernel = "gaussian",     na.rm = T)
#> 
#> Data: y (855 obs.);	Bandwidth 'bw' = 0.0325
#> 
#>        x                  y           
#>  Min.   :-0.05727   Min.   :0.000194  
#>  1st Qu.: 0.17428   1st Qu.:0.016431  
#>  Median : 0.40583   Median :0.776302  
#>  Mean   : 0.40583   Mean   :1.078597  
#>  3rd Qu.: 0.63738   3rd Qu.:1.954287  
#>  Max.   : 0.86893   Max.   :4.004518  
#> 
#> $fits$`1:1`$xy_peaks
#> # A tibble: 6 × 5
#>        x     y counts_per_bin discarded from 
#>    <dbl> <dbl>          <int> <lgl>     <chr>
#> 1 0.1     4                41 FALSE     KDE  
#> 2 0.37    2.56             21 FALSE     KDE  
#> 3 0.0877  3.80             60 FALSE     BMix 
#> 4 0.345   2.43             19 FALSE     BMix 
#> 5 0.425   1.82             24 FALSE     BMix 
#> 6 0.157   2.84             27 FALSE     BMix 
#> 
#> 
#> $fits$`2:0`
#> $fits$`2:0`$matching
#> # A tibble: 2 × 14
#> # Rowwise: 
#>   mutation_multiplic… karyotype  peak     x     y counts_per_bin discarded from 
#>                 <dbl> <chr>     <dbl> <dbl> <dbl>          <int> <lgl>     <chr>
#> 1                   1 2:0       0.367 0.369 0.873             47 FALSE     BMix 
#> 2                   2 2:0       0.734 0.736 5.77             340 FALSE     BMix 
#> # … with 6 more variables: offset_VAF <dbl>, offset <dbl>, weight <dbl>,
#> #   epsilon <dbl>, VAF_tolerance <dbl>, matched <lgl>
#> 
#> $fits$`2:0`$density
#> 
#> Call:
#> 	density.default(x = y, adjust = kernel_adjust, kernel = "gaussian",     na.rm = T)
#> 
#> Data: y (5374 obs.);	Bandwidth 'bw' = 0.01276
#> 
#>        x                  y           
#>  Min.   :0.007179   Min.   :0.000132  
#>  1st Qu.:0.252758   1st Qu.:0.051139  
#>  Median :0.498337   Median :0.324065  
#>  Mean   :0.498337   Mean   :1.017008  
#>  3rd Qu.:0.743916   3rd Qu.:0.928419  
#>  Max.   :0.989495   Max.   :5.776616  
#> 
#> $fits$`2:0`$xy_peaks
#> # A tibble: 6 × 5
#>       x     y counts_per_bin discarded from 
#>   <dbl> <dbl>          <int> <lgl>     <chr>
#> 1 0.09  0.01               2 TRUE      KDE  
#> 2 0.35  0.98              56 FALSE     KDE  
#> 3 0.4   0.89              59 FALSE     KDE  
#> 4 0.74  5.78             340 FALSE     KDE  
#> 5 0.736 5.77             340 FALSE     BMix 
#> 6 0.369 0.873             47 FALSE     BMix 
#> 
#> 
#> $fits$`2:1`
#> $fits$`2:1`$matching
#> # A tibble: 2 × 14
#> # Rowwise: 
#>   mutation_multiplic… karyotype  peak     x     y counts_per_bin discarded from 
#>                 <dbl> <chr>     <dbl> <dbl> <dbl>          <int> <lgl>     <chr>
#> 1                   1 2:1       0.268  0.27  4.16           3717 FALSE     KDE  
#> 2                   2 2:1       0.537  0.54  3.02           2747 FALSE     KDE  
#> # … with 6 more variables: offset_VAF <dbl>, offset <dbl>, weight <dbl>,
#> #   epsilon <dbl>, VAF_tolerance <dbl>, matched <lgl>
#> 
#> $fits$`2:1`$density
#> 
#> Call:
#> 	density.default(x = y, adjust = kernel_adjust, kernel = "gaussian",     na.rm = T)
#> 
#> Data: y (88422 obs.);	Bandwidth 'bw' = 0.01338
#> 
#>        x                  y           
#>  Min.   :-0.02544   Min.   :0.000014  
#>  1st Qu.: 0.21318   1st Qu.:0.018490  
#>  Median : 0.45180   Median :0.508017  
#>  Mean   : 0.45180   Mean   :1.046670  
#>  3rd Qu.: 0.69042   3rd Qu.:1.932413  
#>  Max.   : 0.92904   Max.   :4.159403  
#> 
#> $fits$`2:1`$xy_peaks
#> # A tibble: 6 × 5
#>       x     y counts_per_bin discarded from 
#>   <dbl> <dbl>          <int> <lgl>     <chr>
#> 1 0.27   4.16           3717 FALSE     KDE  
#> 2 0.54   3.02           2747 FALSE     KDE  
#> 3 0.78   0.01             12 TRUE      KDE  
#> 4 0.86   0.01              8 TRUE      KDE  
#> 5 0.266  4.16           3717 FALSE     BMix 
#> 6 0.533  3.02           2640 FALSE     BMix 
#> 
#> 
#> $fits$`2:2`
#> $fits$`2:2`$matching
#> # A tibble: 2 × 14
#> # Rowwise: 
#>   mutation_multiplic… karyotype  peak     x     y counts_per_bin discarded from 
#>                 <dbl> <chr>     <dbl> <dbl> <dbl>          <int> <lgl>     <chr>
#> 1                   1 2:2       0.212 0.22   0.91            237 FALSE     KDE  
#> 2                   2 2:2       0.423 0.425  6.64           1652 FALSE     BMix 
#> # … with 6 more variables: offset_VAF <dbl>, offset <dbl>, weight <dbl>,
#> #   epsilon <dbl>, VAF_tolerance <dbl>, matched <lgl>
#> 
#> $fits$`2:2`$density
#> 
#> Call:
#> 	density.default(x = y, adjust = kernel_adjust, kernel = "gaussian",     na.rm = T)
#> 
#> Data: y (25290 obs.);	Bandwidth 'bw' = 0.007472
#> 
#>        x                    y           
#>  Min.   :-0.0001934   Min.   :0.000053  
#>  1st Qu.: 0.2357220   1st Qu.:0.007649  
#>  Median : 0.4716374   Median :0.104161  
#>  Mean   : 0.4716374   Mean   :1.058694  
#>  3rd Qu.: 0.7075528   3rd Qu.:0.892707  
#>  Max.   : 0.9434683   Max.   :6.669876  
#> 
#> $fits$`2:2`$xy_peaks
#> # A tibble: 11 × 5
#>        x     y counts_per_bin discarded from 
#>    <dbl> <dbl>          <int> <lgl>     <chr>
#>  1 0.22  0.91             237 FALSE     KDE  
#>  2 0.43  6.67            1643 FALSE     KDE  
#>  3 0.66  0.01               0 TRUE      KDE  
#>  4 0.68  0.01               0 TRUE      KDE  
#>  5 0.72  0.01               1 TRUE      KDE  
#>  6 0.77  0.01               1 TRUE      KDE  
#>  7 0.81  0.01               1 TRUE      KDE  
#>  8 0.85  0.01               4 TRUE      KDE  
#>  9 0.88  0.01               6 TRUE      KDE  
#> 10 0.223 0.896            237 FALSE     BMix 
#> 11 0.425 6.64            1652 FALSE     BMix 
#> 
#> 
#> 
#> $matches
#> # A tibble: 7 × 17
#> # Rowwise: 
#>   mutation_multiplic… karyotype  peak     x     y counts_per_bin discarded from 
#>                 <dbl> <chr>     <dbl> <dbl> <dbl>          <int> <lgl>     <chr>
#> 1                   1 1:1       0.367 0.37  2.56              21 FALSE     KDE  
#> 2                   1 2:0       0.367 0.369 0.873             47 FALSE     BMix 
#> 3                   2 2:0       0.734 0.736 5.77             340 FALSE     BMix 
#> 4                   1 2:1       0.268 0.27  4.16            3717 FALSE     KDE  
#> 5                   2 2:1       0.537 0.54  3.02            2747 FALSE     KDE  
#> 6                   1 2:2       0.212 0.22  0.91             237 FALSE     KDE  
#> 7                   2 2:2       0.423 0.425 6.64            1652 FALSE     BMix 
#> # … with 9 more variables: offset_VAF <dbl>, offset <dbl>, weight <dbl>,
#> #   epsilon <dbl>, VAF_tolerance <dbl>, matched <lgl>, score <dbl>,
#> #   purity_error <dbl>, QC <chr>
#> 
#> $matching_strategy
#> [1] "closest"
#> 
#> $purity_error
#> [1] 0.03
#> 
#> $min_karyotype_size
#> [1] 0
#> 
#> $p_binsize_peaks
#> [1] 0.005
#> 
#> $QC
#> [1] "PASS"
#> 
#> $KDE
#> [1] TRUE
#> 
#> $general
#> $general$analysis
#> [1] "4:2" "4:0" "3:2" "3:3" "3:1" "3:0" "4:3"
#> 
#> $general$params
#> $general$params$n_min
#> [1] 150
#> 
#> $general$params$epsilon
#> [1] 0.03
#> 
#> $general$params$n_bootstrap
#> [1] 5
#> 
#> 
#> $general$expected_peaks
#> # A tibble: 24 × 9
#>    minor Major ploidy multiplicity purity  peak karyotype matched n      
#>    <dbl> <dbl>  <dbl>        <int>  <dbl> <dbl> <chr>     <lgl>   <table>
#>  1     2     4      6            1  0.734 0.149 4:2       TRUE     1441  
#>  2     2     4      6            2  0.734 0.297 4:2       TRUE     1441  
#>  3     2     4      6            3  0.734 0.446 4:2       FALSE    1441  
#>  4     2     4      6            4  0.734 0.595 4:2       TRUE     1441  
#>  5     0     4      4            1  0.734 0.212 4:0       TRUE     1752  
#>  6     0     4      4            2  0.734 0.423 4:0       TRUE     1752  
#>  7     0     4      4            3  0.734 0.635 4:0       FALSE    1752  
#>  8     0     4      4            4  0.734 0.847 4:0       TRUE     1752  
#>  9     2     3      5            1  0.734 0.175 3:2       TRUE    58384  
#> 10     2     3      5            2  0.734 0.349 3:2       TRUE    58384  
#> # … with 14 more rows
#> 
#> $general$data_peaks
#> # A tibble: 81 × 7
#>        x     y counts_per_bin discarded from  karyotype n      
#>    <dbl> <dbl>          <int> <lgl>     <chr> <chr>     <table>
#>  1  0.15  0.93             19 FALSE     KDE   4:2       1441   
#>  2  0.3   2.96             59 FALSE     KDE   4:2       1441   
#>  3  0.59  3.14             62 FALSE     KDE   4:2       1441   
#>  4  0.9   0.02              0 TRUE      KDE   4:2       1441   
#>  5  0.16  0.98             19 FALSE     KDE   4:2       1441   
#>  6  0.29  2.75             47 FALSE     KDE   4:2       1441   
#>  7  0.91  0.01              0 TRUE      KDE   4:2       1441   
#>  8  0.42  3.55             88 FALSE     KDE   4:0       1752   
#>  9  0.84  1.86             59 FALSE     KDE   4:0       1752   
#> 10  0.22  0.71             19 FALSE     KDE   4:0       1752   
#> # … with 71 more rows
#> 
#> $general$data_densities
#> # A tibble: 3,584 × 4
#>          x        y karyotype n      
#>      <dbl>    <dbl> <chr>     <table>
#>  1 -0.0844 0.000252 4:2       1441   
#>  2 -0.0822 0.000307 4:2       1441   
#>  3 -0.0800 0.000371 4:2       1441   
#>  4 -0.0778 0.000447 4:2       1441   
#>  5 -0.0757 0.000535 4:2       1441   
#>  6 -0.0735 0.000639 4:2       1441   
#>  7 -0.0713 0.000764 4:2       1441   
#>  8 -0.0692 0.000909 4:2       1441   
#>  9 -0.0670 0.00108  4:2       1441   
#> 10 -0.0648 0.00127  4:2       1441   
#> # … with 3,574 more rows
#> 
#> $general$summary
#> # A tibble: 7 × 5
#> # Groups:   karyotype, matched [7]
#>   karyotype n       matched mismatched  prop
#>   <chr>     <table>   <int>      <dbl> <dbl>
#> 1 3:0       26622         3          0  1   
#> 2 3:1       48704         3          0  1   
#> 3 3:2       58384         3          0  1   
#> 4 3:3       16790         3          0  1   
#> 5 4:0        1752         3          1  0.75
#> 6 4:2        1441         3          1  0.75
#> 7 4:3         359         3          1  0.75
#> 
#> 
#> $subclonal
#> $subclonal$params
#> $subclonal$params$n_min
#> [1] 150
#> 
#> $subclonal$params$epsilon
#> [1] 0.03
#> 
#> $subclonal$params$n_bootstrap
#> [1] 5
#> 
#> 
#> $subclonal$expected_peaks
#> # A tibble: 98 × 16
#>    mutation karyotype_1 genotype_1 karyotype_2 genotype_2 n.clone_1 n.clone_2
#>    <chr>    <chr>       <chr>      <chr>       <chr>          <dbl>     <dbl>
#>  1 LISKJKOW 2:1         A1A2B1     NA          NA                 1         0
#>  2 DFHSKVBB NA          NA         2:2         A1A2B1B2           0         1
#>  3 WOEJJHRA 2:1         A1A2B1     2:2         A1A2B1B2           1         1
#>  4 DFBNKIRY 2:1         A1A2B1     2:2         A1A2B1B2           1         2
#>  5 OXXXPRUK 2:1         A1A2B1     2:2         A1A2B1B2           2         2
#>  6 CSQYNXSZ NA          NA         2:2         A1A2B1B2           0         1
#>  7 HHFSUPHA 2:1         A1A2B1     2:2         A1A2B1B2           1         1
#>  8 MPCFHOYR 2:1         A1A2B1     2:2         A1A2B1B2           1         2
#>  9 HCSVOLTD 2:1         A1A2B1     2:2         A1A2B1B2           2         2
#> 10 AWJHZQOX NA          NA         2:1         A2B1B2             0         1
#> # … with 88 more rows, and 9 more variables: peak <dbl>,
#> #   genotype_initial <chr>, model <chr>, model_id <chr>, role <chr>,
#> #   segment_id <chr>, matched <lgl>, size <chr>, clones <chr>
#> 
#> $subclonal$data_peaks
#> # A tibble: 99 × 8
#>        x     y counts_per_bin discarded from  segment_id    size          clones
#>    <dbl> <dbl>          <int> <lgl>     <chr> <chr>         <chr>         <chr> 
#>  1  0.06 0.07               1 TRUE      KDE   chr6@82432583 (1.8Mb, n = … 2:1 0…
#>  2  0.19 1.07               5 FALSE     KDE   chr6@82432583 (1.8Mb, n = … 2:1 0…
#>  3  0.27 0.99               0 FALSE     KDE   chr6@82432583 (1.8Mb, n = … 2:1 0…
#>  4  0.4  5.69              27 FALSE     KDE   chr6@82432583 (1.8Mb, n = … 2:1 0…
#>  5  0.2  1.04               2 FALSE     KDE   chr6@82432583 (1.8Mb, n = … 2:1 0…
#>  6  0.43 5.51              23 FALSE     KDE   chr6@82432583 (1.8Mb, n = … 2:1 0…
#>  7  0.26 0.983             11 FALSE     KDE   chr6@82432583 (1.8Mb, n = … 2:1 0…
#>  8  0.44 5.36              15 FALSE     KDE   chr6@82432583 (1.8Mb, n = … 2:1 0…
#>  9  0.12 0.163              0 TRUE      KDE   chr6@82432583 (1.8Mb, n = … 2:1 0…
#> 10  0.25 0.939              2 FALSE     KDE   chr6@82432583 (1.8Mb, n = … 2:1 0…
#> # … with 89 more rows
#> 
#> $subclonal$data_densities
#>               x            y    segment_id             size            clones
#> 1   0.001568149 0.0007878617 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 2   0.002780133 0.0009499095 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 3   0.003992116 0.0011386328 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 4   0.005204100 0.0013608318 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 5   0.006416084 0.0016266229 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 6   0.007628067 0.0019332783 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 7   0.008840051 0.0022850377 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 8   0.010052034 0.0026863146 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 9   0.011264018 0.0031469690 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 10  0.012476001 0.0036846438 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 11  0.013687985 0.0042912578 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 12  0.014899968 0.0049720013 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 13  0.016111952 0.0057320714 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 14  0.017323935 0.0065828315 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 15  0.018535919 0.0075501976 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 16  0.019747903 0.0086165883 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 17  0.020959886 0.0097862455 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 18  0.022171870 0.0110630657 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 19  0.023383853 0.0124561190 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 20  0.024595837 0.0139958673 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 21  0.025807820 0.0156523189 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 22  0.027019804 0.0174256462 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 23  0.028231787 0.0193152457 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 24  0.029443771 0.0213230027 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 25  0.030655755 0.0234735254 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 26  0.031867738 0.0257268593 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 27  0.033079722 0.0280761522 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 28  0.034291705 0.0305135464 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 29  0.035503689 0.0330308890 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 30  0.036715672 0.0356310845 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 31  0.037927656 0.0382756503 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 32  0.039139639 0.0409503100 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 33  0.040351623 0.0436400634 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 34  0.041563607 0.0463291940 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 35  0.042775590 0.0489848086 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 36  0.043987574 0.0515887009 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 37  0.045199557 0.0541233371 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 38  0.046411541 0.0565711364 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 39  0.047623524 0.0589146665 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 40  0.048835508 0.0610910090 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 41  0.050047491 0.0631121671 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 42  0.051259475 0.0649679236 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 43  0.052471458 0.0666459865 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 44  0.053683442 0.0681352806 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 45  0.054895426 0.0693668422 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 46  0.056107409 0.0703803752 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 47  0.057319393 0.0711829612 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 48  0.058531376 0.0717747025 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 49  0.059743360 0.0721575497 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 50  0.060955343 0.0722854696 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 51  0.062167327 0.0722132544 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 52  0.063379310 0.0719660991 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 53  0.064591294 0.0715582093 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 54  0.065803278 0.0710053585 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 55  0.067015261 0.0703028140 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 56  0.068227245 0.0695033178 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 57  0.069439228 0.0686389365 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 58  0.070651212 0.0677325062 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 59  0.071863195 0.0668073198 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 60  0.073075179 0.0658979856 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 61  0.074287162 0.0650410966 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 62  0.075499146 0.0642577698 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 63  0.076711129 0.0635695503 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 64  0.077923113 0.0629971244 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 65  0.079135097 0.0625949083 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 66  0.080347080 0.0623774454 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 67  0.081559064 0.0623413742 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 68  0.082771047 0.0624982752 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 69  0.083983031 0.0628581263 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 70  0.085195014 0.0634710735 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 71  0.086406998 0.0643326216 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 72  0.087618981 0.0654144340 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 73  0.088830965 0.0667157115 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 74  0.090042949 0.0682342985 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 75  0.091254932 0.0700017605 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 76  0.092466916 0.0720038919 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 77  0.093678899 0.0742000200 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 78  0.094890883 0.0765819233 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 79  0.096102866 0.0791410524 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 80  0.097314850 0.0818922422 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 81  0.098526833 0.0848236694 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 82  0.099738817 0.0878989719 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 83  0.100950800 0.0911113696 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 84  0.102162784 0.0944548761 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 85  0.103374768 0.0979408230 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 86  0.104586751 0.1015715100 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 87  0.105798735 0.1053233276 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 88  0.107010718 0.1091980133 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 89  0.108222702 0.1131986196 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 90  0.109434685 0.1173462615 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 91  0.110646669 0.1216663589 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 92  0.111858652 0.1261418222 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 93  0.113070636 0.1307842323 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 94  0.114282620 0.1356061994 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 95  0.115494603 0.1406429724 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 96  0.116706587 0.1459480451 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 97  0.117918570 0.1514938965 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 98  0.119130554 0.1572979419 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 99  0.120342537 0.1633779075 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 100 0.121554521 0.1697778342 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 101 0.122766504 0.1765806771 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 102 0.123978488 0.1837314051 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 103 0.125190471 0.1912477355 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 104 0.126402455 0.1991470930 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 105 0.127614439 0.2074730302 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 106 0.128826422 0.2163376660 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 107 0.130038406 0.2256494756 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 108 0.131250389 0.2354228508 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 109 0.132462373 0.2456716219 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 110 0.133674356 0.2564311948 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 111 0.134886340 0.2678419188 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 112 0.136098323 0.2797752767 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 113 0.137310307 0.2922408286 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 114 0.138522291 0.3052474204 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 115 0.139734274 0.3188172796 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 116 0.140946258 0.3331151335 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 117 0.142158241 0.3479791782 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 118 0.143370225 0.3634124649 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 119 0.144582208 0.3794169641 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 120 0.145794192 0.3959974253 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 121 0.147006175 0.4133302381 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 122 0.148218159 0.4312265464 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 123 0.149430143 0.4496790457 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 124 0.150642126 0.4686786614 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 125 0.151854110 0.4882145385 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 126 0.153066093 0.5084275978 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 127 0.154278077 0.5291329817 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 128 0.155490060 0.5503010432 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 129 0.156702044 0.5719055228 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 130 0.157914027 0.5939175450 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 131 0.159126011 0.6163945193 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 132 0.160337994 0.6391798120 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 133 0.161549978 0.6622166215 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 134 0.162761962 0.6854557771 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 135 0.163973945 0.7088449643 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 136 0.165185929 0.7323210811 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 137 0.166397912 0.7557686671 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 138 0.167609896 0.7791142113 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 139 0.168821879 0.8022848198 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 140 0.170033863 0.8252049515 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 141 0.171245846 0.8476701334 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 142 0.172457830 0.8696178835 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 143 0.173669814 0.8909860819 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 144 0.174881797 0.9116879191 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 145 0.176093781 0.9316361193 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 146 0.177305764 0.9505014402 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 147 0.178517748 0.9682829833 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 148 0.179729731 0.9849742658 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 149 0.180941715 1.0004975148 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 150 0.182153698 1.0147781543 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 151 0.183365682 1.0274274755 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 152 0.184577665 1.0385086106 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 153 0.185789649 1.0481141276 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 154 0.187001633 1.0562050011 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 155 0.188213616 1.0627490013 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 156 0.189425600 1.0673965062 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 157 0.190637583 1.0702684129 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 158 0.191849567 1.0715700772 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 159 0.193061550 1.0713227820 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 160 0.194273534 1.0695560122 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 161 0.195485517 1.0660500429 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 162 0.196697501 1.0609694578 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 163 0.197909485 1.0545952164 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 164 0.199121468 1.0470085883 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 165 0.200333452 1.0382971926 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 166 0.201545435 1.0284126633 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 167 0.202757419 1.0175466747 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 168 0.203969402 1.0059742512 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 169 0.205181386 0.9938128420 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 170 0.206393369 0.9811819951 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 171 0.207605353 0.9681825265 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 172 0.208817336 0.9550196713 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 173 0.210029320 0.9418706228 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 174 0.211241304 0.9288531142 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 175 0.212453287 0.9160823085 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 176 0.213665271 0.9037406384 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 177 0.214877254 0.8920360904 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 178 0.216089238 0.8809853254 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 179 0.217301221 0.8706732811 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 180 0.218513205 0.8611788810 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 181 0.219725188 0.8526861575 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 182 0.220937172 0.8453901924 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 183 0.222149156 0.8391371470 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 184 0.223361139 0.8339585853 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 185 0.224573123 0.8298786074 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 186 0.225785106 0.8270190870 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 187 0.226997090 0.8255343980 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 188 0.228209073 0.8251458972 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 189 0.229421057 0.8258278111 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 190 0.230633040 0.8275477604 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 191 0.231845024 0.8303363769 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 192 0.233057008 0.8342715833 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 193 0.234268991 0.8390385074 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 194 0.235480975 0.8445681624 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 195 0.236692958 0.8507880603 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 196 0.237904942 0.8576503008 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 197 0.239116925 0.8651313842 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 198 0.240328909 0.8729835820 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 199 0.241540892 0.8811236772 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#> 200 0.242752876 0.8894692154 chr6@82432583 (1.8Mb, n = 301) 2:1 0.19 2:2 0.81
#>  [ reached 'max' / getOption("max.print") -- omitted 3384 rows ]
#> 
#> $subclonal$mutations
#> # A tibble: 18,614 × 58
#>    chr       from       to ref   alt   sample         icgc_donor_id project_code
#>    <chr>    <dbl>    <dbl> <chr> <chr> <chr>          <chr>         <chr>       
#>  1 chr6  82434981 82434981 G     A     51893d3f-e7f3… DO220903      MELA-AU     
#>  2 chr6  82445298 82445298 C     T     51893d3f-e7f3… DO220903      MELA-AU     
#>  3 chr6  82479209 82479209 G     T     51893d3f-e7f3… DO220903      MELA-AU     
#>  4 chr6  82486512 82486512 G     A     51893d3f-e7f3… DO220903      MELA-AU     
#>  5 chr6  82496979 82496979 C     T     51893d3f-e7f3… DO220903      MELA-AU     
#>  6 chr6  82516071 82516071 G     A     51893d3f-e7f3… DO220903      MELA-AU     
#>  7 chr6  82522451 82522451 G     A     51893d3f-e7f3… DO220903      MELA-AU     
#>  8 chr6  82534035 82534035 C     T     51893d3f-e7f3… DO220903      MELA-AU     
#>  9 chr6  82537650 82537650 C     T     51893d3f-e7f3… DO220903      MELA-AU     
#> 10 chr6  82572477 82572477 T     C     51893d3f-e7f3… DO220903      MELA-AU     
#> # … with 18,604 more rows, and 50 more variables: icgc_specimen_id <chr>,
#> #   icgc_sample_id <chr>, submitted_sample_id <chr>,
#> #   submitted_matched_sample_id <chr>, chromosome_strand <dbl>,
#> #   assembly_version <chr>, mutation_type <chr>, reference_genome_allele <chr>,
#> #   quality_score <lgl>, probability <lgl>, verification_status <chr>,
#> #   verification_platform <lgl>, biological_validation_status <chr>,
#> #   biological_validation_platform <lgl>, gene_build_version <dbl>, …
#> 
#> $subclonal$summary
#> # A tibble: 15 × 6
#>    segment_id     model_id                     model      prop size       clones
#>    <chr>          <chr>                        <chr>     <dbl> <chr>      <chr> 
#>  1 chr6@82432583  A1A2B1 -> A1A2B1 -> A1A2B1B2 linear     1    (1.8Mb, n… 2:1 0…
#>  2 chr6@82432583  A1A2B1 -> A1A2B1 | A1A2B1B2  branching  1    (1.8Mb, n… 2:1 0…
#>  3 chr6@82432583  A1A2B1 -> A1A2B1B2 -> A2B1B2 linear     1    (1.8Mb, n… 2:1 0…
#>  4 chr11@202253   A1A2B1 -> A1A2B1 -> A1A2B1B2 linear     1    (5.2Mb, n… 2:1 0…
#>  5 chr11@202253   A1A2B1 -> A1A2B1 | A1A2B1B2  branching  1    (5.2Mb, n… 2:1 0…
#>  6 chr11@202253   A1A2B1 -> A1A2B1B2 -> A2B1B2 linear     1    (5.2Mb, n… 2:1 0…
#>  7 chr11@5372292  A1A2B1 -> A1A2B1 -> A1A2B1B2 linear     1    (12Mb, n … 2:1 0…
#>  8 chr11@17365005 A1A2B1 -> A1A2B1 -> A1A2B1B2 linear     1    (31.6Mb, … 2:1 0…
#>  9 chr11@48918601 A1A2B1 -> A1A2B1 -> A1A2B1B2 linear     1    (2.7Mb, n… 2:1 0…
#> 10 chr11@48918601 A1A2B1 -> A1A2B1 | A1A2B1B2  branching  1    (2.7Mb, n… 2:1 0…
#> 11 chr11@48918601 A1A2B1 -> A1A2B1B2 -> A2B1B2 linear     1    (2.7Mb, n… 2:1 0…
#> 12 chr11@51600000 A1A2B1 -> A1A2B1 -> A1A2B1B2 linear     0.75 (4.1Mb, n… 2:1 0…
#> 13 chr11@55700000 A1A2B1 -> A1A2B1 -> A1A2B1B2 linear     1    (78.8Mb, … 2:1 0…
#> 14 chr11@55700000 A1A2B1 -> A1A2B1 | A1A2B1B2  branching  1    (78.8Mb, … 2:1 0…
#> 15 chr11@55700000 A1A2B1 -> A1A2B1B2 -> A2B1B2 linear     1    (78.8Mb, … 2:1 0…
#> 
#> 

x = analyze_peaks(x, matching_strategy = "rightmost")
#> 
#> ℹ Predicting coding
#> ── Peak analysis: simple CNAs ──────────────────────────────────────────────────
#> ℹ Predicting coding
#> 
#> ℹ Analysing 120065 mutations mapping to karyotype(s) 2:1, 2:2, 2:0, 1:1, and 1:0.
#> ℹ Predicting coding
#> ℹ Mixed type peak detection for karyotype 1:0 (124 mutations)
#> ℹ Predicting coding
#> ℹ Mixed type peak detection for karyotype 1:1 (855 mutations)
#> ℹ Predicting coding
#> ℹ Mixed type peak detection for karyotype 2:0 (5374 mutations)
#> ℹ Predicting coding
#> ℹ Mixed type peak detection for karyotype 2:1 (88422 mutations)
#> ℹ Predicting coding
#> ℹ Mixed type peak detection for karyotype 2:2 (25290 mutations)
#> ℹ Predicting coding
#> # A tibble: 8 × 16
#> # Rowwise: 
#>   mutation_multiplicity karyotype  peak delta_vaf     x     y counts_per_bin
#>                   <dbl> <chr>     <dbl>     <dbl> <dbl> <dbl>          <int>
#> 1                     1 2:1       0.268   0.0134  0.266 4.14            3720
#> 2                     2 2:1       0.537   0.0268  0.54  3               2729
#> 3                     1 2:2       0.212   0.00831 0.21  0.93             198
#> 4                     2 2:2       0.423   0.0166  0.421 6.85            1732
#> 5                     1 2:0       0.367   0.025   0.359 0.864             50
#> 6                     2 2:0       0.734   0.05    0.735 5.43             265
#> 7                     1 1:1       0.367   0.025   0.368 2.42              15
#> 8                     1 1:0       0.580   0.0624  0.570 4.16               7
#> # … with 9 more variables: discarded <lgl>, from <chr>, offset_VAF <dbl>,
#> #   offset <dbl>, weight <dbl>, epsilon <dbl>, VAF_tolerance <dbl>,
#> #   matched <lgl>, QC <chr>
#> ✔ Peak detection PASS with r = 0.00711169452392766 - maximum purity error ε = 0.05.
#> ℹ Predicting coding
#> 
#> ℹ Predicting coding
#> ── Peak analysis: complex CNAs ─────────────────────────────────────────────────
#> ℹ Predicting coding
#> 
#> ℹ Karyotypes 3:2, 3:1, 3:0, 3:3, 4:0, 4:2, 4:3, 5:2, 5:3, and 6:3 with >100 mutation(s). Using epsilon = 0.05.
#> ℹ Predicting coding
#> Error in arrange(., prop %>% desc): Problem with the implicit `transmute()` step.
#> ✖ Problem while computing `..1 = prop %>% desc`.
#> Caused by error:
#> ! unable to find an inherited method for function ‘desc’ for signature ‘"numeric"’
print(x)
#> ── [ CNAqc ]  293736 mutations in 667 segments (654 clonal, 13 subclonal). Genom
#> ℹ Predicting coding
#> 
#> ℹ Predicting coding
#> ── Clonal CNAs 
#> ℹ Predicting coding
#> 
#>   2:1  [n = 88422, L =   692 Mb] ■■■■■■■■■■■■■■■■■■■■■■■■■■■
#>   3:2  [n = 58384, L =   417 Mb] ■■■■■■■■■■■■■■■■■■  { BRAF }
#>   3:1  [n = 48704, L =   380 Mb] ■■■■■■■■■■■■■■■
#>   3:0  [n = 26622, L =   360 Mb] ■■■■■■■■  { CDKN2A }
#>   2:2  [n = 25290, L =   253 Mb] ■■■■■■■■
#>   3:3  [n = 16790, L =   115 Mb] ■■■■■
#>   2:0  [n =  5374, L =    67 Mb] ■■
#>   4:0  [n =  1752, L =    22 Mb] ■  { TP53 }
#>   4:2  [n =  1441, L =    11 Mb] 
#>   1:1  [n =   855, L =     9 Mb] 
#> 
#> 
#> ℹ Predicting coding
#> ── Subclonal CNAs (showing up to 10 segments) 
#> ℹ Predicting coding
#> 
#>  chr11@55700000  [n =  10468, L =  78.75 Mb]   2:1 (0.21)   2:2 (0.79) ■■■■■■■■■■
#>  chr11@17365005  [n =   5389, L =  31.55 Mb]   2:1 (0.21)   2:2 (0.79) ■■■■■
#>   chr11@5372292  [n =   1014, L =  11.99 Mb]   2:1 (0.21)   2:2 (0.79) 
#>    chr11@202253  [n =    610, L =   5.17 Mb]   2:1 (0.22)   2:2 (0.78) 
#>  chr11@48918601  [n =    542, L =   2.68 Mb]   2:1 (0.25)   2:2 (0.75) 
#>   chr6@82432583  [n =    301, L =   1.81 Mb]   2:1 (0.19)   2:2 (0.81) 
#>  chr11@51600000  [n =    290, L =    4.1 Mb]   2:1 (0.26)   2:2 (0.74) 
#>   chr6@81896364  [n =     69, L =   0.54 Mb]   2:1 (0.19)   2:2 (0.81) 
#>   chr6@93956180  [n =     41, L =   0.11 Mb]    2:1 (0.2)    2:2 (0.8) 
#>   chr8@42633277  [n =     13, L =   0.26 Mb]   2:1 (0.28)   2:2 (0.72) 
#> 
#> ℹ Sample Purity: 73.4% ~ Ploidy: 3.
#> ℹ Predicting coding
#> 
#> ℹ There are 3 annotated driver(s) mapped to clonal CNAs.
#> ℹ Predicting coding
#>          chr      from        to ref alt DP NV       VAF driver_label is_driver
#>        chr17   7577082   7577082   C   T 78 70 0.8974359         TP53      TRUE
#>         chr7 140453136 140453136   A   T 95 54 0.5684211         BRAF      TRUE
#>         chr9  21971120  21971120   G   A 23 14 0.6086957       CDKN2A      TRUE
#> 
#> ℹ Predicting coding
#> ──  PASS  Peaks QC closest: 100%, λ = -0.0104. Purity correction: -1%. ─────────
#> ℹ Predicting coding
#> 
#> ℹ 1:1 ~ n = 855   (  1%) →  PASS  -0.006 
#> ℹ Predicting coding
#> ℹ 2:0 ~ n = 5374  (  4%) →  PASS  -0.003   PASS  -0.002 
#> ℹ Predicting coding
#> ℹ 2:1 ~ n = 88422 ( 74%) →  PASS  -0.006   PASS  -0.006 
#> ℹ Predicting coding
#> ℹ 2:2 ~ n = 25290 ( 21%) →  PASS  -0.053   PASS  -0.004 
#> ℹ Predicting coding
#> 
#> ℹ Predicting coding
#> ── General peak QC (154052 mutations):  PASS  21  FAIL  3 - epsilon = 0.03. ────
#> ℹ Predicting coding
#> 
#> ℹ 3:0 ~ n = 26622 ( 17%) →  PASS  3  FAIL  0
#> ℹ Predicting coding
#> ℹ 3:1 ~ n = 48704 ( 32%) →  PASS  3  FAIL  0
#> ℹ Predicting coding
#> ℹ 3:2 ~ n = 58384 ( 38%) →  PASS  3  FAIL  0
#> ℹ Predicting coding
#> ℹ 3:3 ~ n = 16790 ( 11%) →  PASS  3  FAIL  0
#> ℹ Predicting coding
#> ℹ 4:0 ~ n = 1752  (  1%) →  PASS  3  FAIL  1
#> ℹ Predicting coding
#> ℹ 4:2 ~ n = 1441  (  1%) →  PASS  3  FAIL  1
#> ℹ Predicting coding
#> ℹ 4:3 ~ n = 359   (  0%) →  PASS  3  FAIL  1
#> ℹ Predicting coding
#> Error in arrange(., BR %>% desc, LI %>% desc): Problem with the implicit `transmute()` step.
#> ✖ Problem while computing `..1 = BR %>% desc`.
#> Caused by error:
#> ! unable to find an inherited method for function ‘desc’ for signature ‘"integer"’