Import VEP annotations.

This function imports Ensembl Variant Effect Predictor (VEP) annotations into a CNAqc object, or to a mutations dataframe in the format ready for CNAqc. Therefore the expected uses are

"[Option 1]"Mutations + CNA + Purity -> CNAqc -> VEP annotation

"[Option 2]"Mutations -> VEP annotation -> Mutations (with VEP) + CNA + Purity -> CNAqc

VEP annotations should have been created using the scripts discussed at the CNAqc webpage utility. At this point, if one has created the input data for CNAqc from the original VCF file, the VEP can be added to the CNAqc object (Option 1), or to the available mutations before creating the CNAqc obejct (Option 2).

VEP annotations and input mutations are matched by exact genomic coordinates.

augment_with_vep(x, vep, columns = NULL)

Arguments

x

A CNAqc object, or a dataframe of mutations in the input format for CNAqc.

maf

The file MAF associated containing the annotations in MAF format.

Value

It depends on `x`

"[Option 1]" A CNAqc object like `x` where the mutations are associated to the MAF annotations, if `x` is a CNAqc object. In this case the S3 print method for `x` will report the presence of MAF annotations.

"[Option 2]" If `x` is a dataframe, the same dataframe augmented with MAF annotations.

See also

Examples

# Example with a CNAqc input object
if(FALSE)
{
   # Create your CNAqc object (omissis here) from an original "file.vcf"
   x = init(mutations = ..., cna = ..., purity = ...)

   # Offline, create your MAF annotations as file "file_vcf.maf" from "file.vcf"
   # vcf2maf file.vcf .... file_vcf.maf

   # Import into R/CNAqc
   x = augment_with_maf(x, maf = "file_vcf.maf")

   # check they are in (there should be many columns with "MAF." prefix)
   x %>% Mutations %>% colnames
}