Determines fragmentation of copy number segments

This functions determines if, at the arm level, the segments are over-fragmented using the statistical test described in the package manual.

detect_arm_overfragmentation(
  x,
  alpha = 0.01,
  genome_percentage_cutoff = 0.2,
  minimum_segments_for_testing = 10
)

Arguments

x
An object of class cnaqc, created by the init function.

alpha
Confidence level for the tests, for instance 0.05.

genome_percentage_cutoff
Segments are considered long or short depending on whether they are longer (in basepairs) than genome_percentage_cutoff * L bases, where L is the arm length for the reference genome. Default is 0.2 (twenty percent).

minimum_segments_for_testing
Smallest number of segments required to actually test a certain arm Default is 10 segments. This number influences the correction for mulitple hypothesis testing.

Value

An object of class cnaqc with the results.

Examples

data('example_dataset_CNAqc')
x = init(example_dataset_CNAqc$snvs, example_dataset_CNAqc$cna, example_dataset_CNAqc$purity)

#>  [ CNAqc - CNA Quality Check ] 
#> 

#> ℹ Using reference genome coordinates for: GRCh38.

#> ℹ Drivers are annotated, but 'gene' column is missing, using mutation location.

#> ! Missing CCF column from CNA calls, adding CCF = 1 assuming clonal CNA calls.

#> ℹ Input n = 12963 mutations for 267 CNA segments (267 clonal, 0 subclonal)

#> 

#> ✔ Mapped n = 12963 mutations to clonal segments (100% of input)

x = detect_arm_overfragmentation(x)

#> ℹ One-tailed Binomial test: 8 tests, alpha 0.01. Short segments: 0.2% of the reference arm.

#> ✔ chr7p,  p = 1.71798691840001e-24 ~ 34 segments, 34 short.

#> ✔ chr1p,  p = 1.62738995200002e-15 ~ 24 segments, 23 short.

#> ✔ chr11q,  p = 2.4788992e-07 ~ 14 segments, 12 short.

#> ✔ chr1q,  p = 4.34176000000001e-08 ~ 13 segments, 12 short.

#> ✔ chr12q,  p = 4.34176000000001e-08 ~ 13 segments, 12 short.

#> ✔ chr3q,  p = 4.52608e-06 ~ 12 segments, 10 short.

#> ✔ chr7q,  p = 4.52608e-06 ~ 12 segments, 10 short.

#> ✔ chr8p,  p = 9.21599999999998e-07 ~ 11 segments, 10 short.

#> ℹ 8 significantly overfragmented chromosome arms (alpha level 0.01).

print(x)

#> $reference_genome
#> [1] "GRCh38"
#> 
#> $snvs
#> # A tibble: 12,963 x 17
#>    chr      from      to ref   alt   FILTER    DP    NV    VAF ANNOVAR_FUNCTION
#>    <chr>   <dbl>   <dbl> <chr> <chr> <chr>  <dbl> <dbl>  <dbl> <chr>           
#>  1 chr1  1027104 1027105 T     G     PASS      60     6 0.1    UTR5            
#>  2 chr1  2248588 2248589 A     C     PASS     127     9 0.0709 intergenic      
#>  3 chr1  2461999 2462000 G     A     PASS     156    65 0.417  upstream        
#>  4 chr1  2727935 2727936 T     C     PASS     180    90 0.5    downstream      
#>  5 chr1  2763397 2763398 C     T     PASS     183    61 0.333  intergenic      
#>  6 chr1  2768208 2768209 C     T     PASS     203   130 0.640  intergenic      
#>  7 chr1  2935590 2935591 C     T     PASS     228   132 0.579  intergenic      
#>  8 chr1  2980032 2980033 C     T     PASS     196    85 0.434  ncRNA_exonic    
#>  9 chr1  3387161 3387162 T     G     PASS     124     6 0.0484 intronic        
#> 10 chr1  3502517 3502518 G     A     PASS      88    10 0.114  intronic        
#> # … with 12,953 more rows, and 7 more variables: GENE <chr>, is_driver <lgl>,
#> #   driver_label <chr>, gene <chr>, type <chr>, karyotype <chr>,
#> #   segment_id <chr>
#> 
#> $cna
#> # A tibble: 267 x 10
#>    chr      from      to length covRatio Major minor   CCF segment_id          n
#>    <chr>   <int>   <int>  <int>    <dbl> <dbl> <dbl> <dbl> <chr>           <int>
#>  1 chr1   8.40e5  1.69e6 8.50e5     1.19     3     2     1 chr1:840009:16…     1
#>  2 chr1   1.69e6  1.82e6 1.25e5     1.26     3     2     1 chr1:1689988:1…    NA
#>  3 chr1   1.82e6  9.80e6 7.98e6     1.19     3     2     1 chr1:1815016:9…    28
#>  4 chr1   1.05e7  1.21e7 1.60e6     1.19     3     2     1 chr1:10479910:…     1
#>  5 chr1   1.21e7  1.22e7 7.51e4     1.24     3     2     1 chr1:12079917:…    NA
#>  6 chr1   1.22e7  1.28e7 6.85e5     1.19     3     2     1 chr1:12154981:…    NA
#>  7 chr1   1.38e7  1.78e7 4.01e6     1.19     3     2     1 chr1:13780016:…     7
#>  8 chr1   1.78e7  2.11e7 3.23e6     1.19     3     2     1 chr1:17849962:…    12
#>  9 chr1   2.11e7  2.16e7 4.80e5     1.26     3     2     1 chr1:21080068:…    NA
#> 10 chr1   2.16e7  2.48e7 3.27e6     1.19     3     2     1 chr1:21559998:…    11
#> # … with 257 more rows
#> 
#> $n_snvs
#> [1] 12963
#> 
#> $n_cna
#> [1] 267
#> 
#> $n_cna_clonal
#> [1] 267
#> 
#> $n_cna_sbclonal
#> [1] 0
#> 
#> $n_karyotype
#> 
#>   2:2   4:2   3:2   2:1   3:0   2:0  16:2  25:2   3:1 106:1  26:2  99:1 
#>  7478  1893  1625  1563   312    81     4     2     2     1     1     1 
#> 
#> $purity
#> [1] 0.89
#> 
#> $l_karyotype
#>        2:2        2:1        3:2        4:2        3:0        2:0        3:1 
#> 1482910013  420405440  357172683  331457138  137498772   39372558     527401 
#>       25:2      106:1        5:0       69:2       26:2       16:2       70:2 
#>     517926     225714     174002     172894     162913     130712      80103 
#>       99:1 
#>      75782 
#> 
#> $ploidy
#> [1] 4
#> 
#> $most_prevalent_karyotype
#> [1] "2:2"
#> 
#> $basepairs_by_karyotype
#> # A tibble: 15 x 4
#>    minor Major          n karyotype
#>    <dbl> <dbl>      <int> <chr>    
#>  1     2     2 1482910013 2:2      
#>  2     1     2  420405440 2:1      
#>  3     2     3  357172683 3:2      
#>  4     2     4  331457138 4:2      
#>  5     0     3  137498772 3:0      
#>  6     0     2   39372558 2:0      
#>  7     1     3     527401 3:1      
#>  8     2    25     517926 25:2     
#>  9     1   106     225714 106:1    
#> 10     0     5     174002 5:0      
#> 11     2    69     172894 69:2     
#> 12     2    26     162913 26:2     
#> 13     2    16     130712 16:2     
#> 14     2    70      80103 70:2     
#> 15     1    99      75782 99:1     
#> 
#> $most_mutations_karyotype
#> [1] "2:2"
#> 
#> $arm_fragmentation
#> $arm_fragmentation$table
#> # A tibble: 41 x 10
#>    chr   arm   n_long n_short jumps total_segments testable  p_value
#>    <chr> <chr>  <dbl>   <dbl> <dbl>          <dbl> <lgl>       <dbl>
#>  1 chr7  p          0      34   206             34 TRUE     1.72e-24
#>  2 chr1  p          1      23     0             24 TRUE     1.63e-15
#>  3 chr1  q          1      12     5             13 TRUE     4.34e- 8
#>  4 chr12 q          1      12   414             13 TRUE     4.34e- 8
#>  5 chr11 q          2      12     0             14 TRUE     2.48e- 7
#>  6 chr8  p          1      10     0             11 TRUE     9.22e- 7
#>  7 chr3  q          2      10     2             12 TRUE     4.53e- 6
#>  8 chr7  q          2      10     3             12 TRUE     4.53e- 6
#>  9 chr11 p          1       8     0              9 FALSE    1.00e+ 0
#> 10 chr15 q          2       7     4              9 FALSE    1.00e+ 0
#> # … with 31 more rows, and 2 more variables: Bonferroni_cutoff <dbl>,
#> #   significant <lgl>
#> 
#> $arm_fragmentation$alpha
#> [1] 0.01
#> 
#> $arm_fragmentation$N_tests
#> [1] 8
#> 
#> $arm_fragmentation$genome_percentage_cutoff
#> [1] 0.2
#> 
#> $arm_fragmentation$minimum_segments_for_testing
#> [1] 10
#> 
#> 
#> attr(,"class")
#> [1] "cnaqc"