Create a CNAqc object.

Creates a CNAqc object from a set of mutation and CNA calls, and a tumour purity value. The resulting object contains the mutations mapped into CNA segments, and allows for the computation of the QC metrics available in the CNAqc package. CNAqc supports `hg19` coordinates.

init(snvs, cna, purity)

Arguments

snvs
A data.frame of mutations with the following fields available: `chr`, `from`, `to` as `hg19` chromosome coordinates. `ref` and `alt` for the reference and alternative alleles, and `DP`, `NV` and `VAF` for the depth (total number of reads with both reference and alternative), the number of variants with the alternative allele and the allele frequency (VAF = NV/DP). Chromosome names must be in the format `chr1`, `chr2`, etc.; alleles should be characters and all other fields numeric.

cna
A data.frame of CNA with the following fields available: `chr`, `from`, `to` as `hg19` chromosome coordinates. `Major` and `minor` for the number of copies of the major allele, and the minor (B-allele). A `CCF` column can be used as a real-value in between 0 and 1 to represent Cancer Cell Fractions; if this is not available `CCF = 1` is set and all calls will refer to clonal segments. Otherwise, segments with `CCF<1` would be considered subclonal CNAs.

purity
Value in between `0` and `1` to represent the proportion of actual tumour content (sometimes called "cellularity").

Value

A CNAqc object of class `cnaqc`, with S3 methods for printing, plotting and analyzing data.

Examples

data('example_dataset_CNAqc', package = 'CNAqc')
print(example_dataset_CNAqc)

#> $snvs
#> # A tibble: 13,141 x 51
#>    chr   from  ID    ref   alt   QUAL  FILTER INFO  id       DP    NV    VAF
#>    <chr> <chr> <chr> <chr> <chr> <chr> <chr>  <chr> <chr> <dbl> <dbl>  <dbl>
#>  1 chr1  1027… NA    T     G     1.73… PASS   DP=8… chr1…    60     6 0.1   
#>  2 chr1  2248… NA    A     C     -0    PASS   DP=1… chr1…   127     9 0.0709
#>  3 chr1  2461… NA    G     A     225   PASS   DP=1… chr1…   156    65 0.417 
#>  4 chr1  2727… NA    T     C     225   PASS   DP=1… chr1…   180    90 0.5   
#>  5 chr1  2763… NA    C     T     225   PASS   DP=1… chr1…   183    61 0.333 
#>  6 chr1  2768… rs18… C     T     225   PASS   DP=2… chr1…   203   130 0.640 
#>  7 chr1  2935… NA    C     T     225   PASS   DP=2… chr1…   228   132 0.579 
#>  8 chr1  2980… NA    C     T     225   PASS   DP=2… chr1…   196    85 0.434 
#>  9 chr1  3387… NA    T     G     -0    PASS   DP=1… chr1…   124     6 0.0484
#> 10 chr1  3502… rs11… G     A     0.00… PASS   DP=9… chr1…    88    10 0.114 
#> # … with 13,131 more rows, and 39 more variables:
#> #   `/icgc/dkfzlsdf/project/hipo/hipo_043/sequencing/whole_genome_sequencing/view-by-pid/H043-W99H5K/tumor/paired/merged-alignment/.merging_0/tumor_H043-W99H5K_merged.mdup.bam` <chr>,
#> #   SEQUENCE_CONTEXT <chr>,
#> #   `INFO_control(VAF=variant_allele_fraction;TSR=total_variant_supporting_reads_incl_lowqual)` <chr>,
#> #   ANNOTATION_control <chr>, DBSNP <chr>, `1K_GENOMES` <chr>,
#> #   ANNOVAR_FUNCTION <chr>, GENE <chr>, EXONIC_CLASSIFICATION <chr>,
#> #   ANNOVAR_TRANSCRIPTS <chr>, SEGDUP <chr>, CYTOBAND <chr>,
#> #   REPEAT_MASKER <chr>, DAC_BLACKLIST <chr>, DUKE_EXCLUDED <chr>,
#> #   HISEQDEPTH <chr>, SELFCHAIN <chr>, MAPABILITY <chr>,
#> #   SIMPLE_TANDEMREPEATS <chr>, CONFIDENCE <chr>, RECLASSIFICATION <chr>,
#> #   PENALTIES <chr>, seqBiasPresent_1 <chr>, seqingBiasPresent_1 <chr>,
#> #   seqBiasPresent_2 <chr>, seqingBiasPresent_2 <chr>, Enhancers <chr>,
#> #   CpGislands <chr>, TFBScons <chr>, ENCODE_DNASE <chr>, miRNAs_snoRNAs <chr>,
#> #   miRBase18 <chr>, COSMIC <chr>, miRNAtargets <chr>, CgiMountains <chr>,
#> #   phastConsElem20bp <chr>, ENCODE_TFBS <chr>, blacklist <chr>, to <dbl>
#> 
#> $cna
#> # A tibble: 482 x 19
#>    chr     from     to length covRatio   TCN SV.Type c1Mean c2Mean  dhEst dhSNPs
#>    <chr>  <int>  <int>  <int>    <dbl> <dbl> <fct>    <dbl>  <dbl>  <dbl>  <dbl>
#>  1 chr1  8.40e5 1.69e6 8.50e5     1.19  5.05 NA        2.02   3.03  0.199  0.192
#>  2 chr1  1.69e6 1.82e6 1.25e5     1.26  5.38 NA        2.19   3.19  0.185  0.182
#>  3 chr1  1.82e6 9.80e6 7.98e6     1.19  5.05 NA        2.02   3.03  0.199  0.190
#>  4 chr1  9.80e6 9.87e6 7.50e4     1.26  5.38 NA       NA     NA    NA     NA    
#>  5 chr1  9.87e6 1.05e7 6.05e5     1.24  5.29 NA       NA     NA    NA      0.228
#>  6 chr1  1.05e7 1.21e7 1.60e6     1.19  5.05 NA        2.02   3.03  0.199  0.181
#>  7 chr1  1.21e7 1.22e7 7.51e4     1.24  5.29 NA        2.04   3.25  0.229  0.207
#>  8 chr1  1.22e7 1.28e7 6.85e5     1.19  5.05 NA        2.02   3.03  0.199  0.200
#>  9 chr1  1.38e7 1.78e7 4.01e6     1.19  5.05 NA        2.02   3.03  0.199  0.193
#> 10 chr1  1.78e7 1.78e7 5.99e4     1.26  5.38 NA       NA     NA    NA      0.158
#> # … with 472 more rows, and 8 more variables: Major <chr>, minor <chr>,
#> #   CNA.type <fct>, NbrOfHetsSNPs <int>, minStart <int>, maxStart <int>,
#> #   minStop <int>, maxStop <int>
#> 
#> $purity
#> [1] 0.89
#> 

x = init(example_dataset_CNAqc$snvs, example_dataset_CNAqc$cna,example_dataset_CNAqc$purity)

#>  [ CNAqc - CNA Quality Check ] 
#> 

#> Warning: [CNAqc] Enforcing numeric for values: 1027104, 2248588, 2461999, 2727935, 2763397, 2768208, ...

#> [CNAqc] Missing CCF column from CNA calls, adding CCF = 1 assuming all calls to be clonal.

#> Warning: [CNAqc] Enforcing numeric for values: 3, 3, 3, NA, NA, 3, ...

#> Warning: NAs introduced by coercion

#> Warning: [CNAqc] Enforcing numeric for values: 2, 2, 2, NA, NA, 2, ...

#> Warning: NAs introduced by coercion

#> 
#> Input  n = 13141 mutations for 482 CNA segments (482 clonal, 0 subclonal)
#> 
#>  Mapping mutations to clonal CNA.
#> 
#> Mapping. n = 13050 mutations mapped to segments (~99% of input) 
#> [CNAqc] The following segments have Major/ minor alleles in non-numeric format or NAs
#> # A tibble: 215 x 20
#>    chr     from     to length covRatio   TCN SV.Type c1Mean c2Mean dhEst dhSNPs
#>    <chr>  <int>  <int>  <int>    <dbl> <dbl> <fct>    <dbl>  <dbl> <dbl>  <dbl>
#>  1 chr1  9.80e6 9.87e6  75009     1.26  5.38 NA          NA     NA    NA NA    
#>  2 chr1  9.87e6 1.05e7 604933     1.24  5.29 NA          NA     NA    NA  0.228
#>  3 chr1  1.78e7 1.78e7  59936     1.26  5.38 NA          NA     NA    NA  0.158
#>  4 chr1  3.09e7 3.09e7  14969     1.24  5.29 NA          NA     NA    NA  0.229
#>  5 chr1  3.14e7 3.16e7 179994     1.24  5.29 NA          NA     NA    NA  0.102
#>  6 chr1  3.45e7 3.45e7  20049     1.26  5.38 NA          NA     NA    NA NA    
#>  7 chr1  5.04e7 5.04e7  44976     1.24  5.29 NA          NA     NA    NA  0.216
#>  8 chr1  5.19e7 5.19e7  24947     1.27  5.43 NA          NA     NA    NA NA    
#>  9 chr1  6.13e7 6.13e7  39974     1.24  5.29 NA          NA     NA    NA  0.233
#> 10 chr1  6.84e7 6.84e7  20002     1.24  5.29 NA          NA     NA    NA  0.235
#> # … with 205 more rows, and 9 more variables: Major <dbl>, minor <dbl>,
#> #   CNA.type <fct>, NbrOfHetsSNPs <int>, minStart <int>, maxStart <int>,
#> #   minStop <int>, maxStop <int>, CCF <dbl>
#> These CNA segments will be removed.
#> 

print(x)

#>       CNAqc  n = 13141 mutations for 267 CNA segments (267 clonal, 0 subclonal) 
#>      Purity  89 cellularity 
#>  Karyotypes  7478 (2:2); 1893 (4:2); 1625 (3:2); 1563 (2:1); 312 (3:0); 81 (2:0); 67 (NA:NA); 16 (NA:2); 4 (104:NA); 4 (16:2); 2 (25:2); 2 (3:1); 1 (106:1); 1 (26:2); 1 (99:1) 
#>  
#>    Peaks QC  FALSE 
#>