Plot CNA segments.

Plot CNA segments as Major/ minor allele, annotating clonal and subclonal CNA calls in two different sets of colors.

plot_segments(
  x,
  chromosomes = paste0("chr", c(1:22, "X", "Y")),
  max_Y_height = 6,
  circular = FALSE,
  cn = "absolute",
  highlight = x$most_prevalent_karyotype,
  ...
)

Arguments

x

An object of class cnaqc, created by the init function.

chromosomes

The chromosome to use for this plot.

max_Y_height

Maximum height for the Y-axis of the plot. Segments witht total copy number (Major plus minor) above max_Y_height are not showm; in that case a point annotation is put on top of the plot, and max_Y_height is annotated in the grid.

circular

Uses a circular layout in polar coordinates to make the segments look like a circos plot. This visualisation can save space.

cn

Type of copy number segment to show on the plot. Either "absolute" for Major and minor annotation, or "total" for the total (Major plus minor) annotation.

highlight

A list of karyotype ids in "Major:minor" notation (e.g., "1:1", "2,1", ...) that will be shadowed with a transparent area. By default, it plots the most prevalent karyotype,

Value

A ggplot object.

Examples

data('example_dataset_CNAqc', package = 'CNAqc')
x = init(example_dataset_CNAqc$mutations, example_dataset_CNAqc$cna,example_dataset_CNAqc$purity)
#>  [ CNAqc - CNA Quality Check ] 
#> 
#> Error in init(example_dataset_CNAqc$mutations, example_dataset_CNAqc$cna,     example_dataset_CNAqc$purity): Avoid using `snvs` if you use `mutations

plot_segments(x)
#> Error in stopifnot(inherits(x, "cnaqc")): object 'x' not found
plot_segments(x, chromosomes = 'chr13')
#> Error in stopifnot(inherits(x, "cnaqc")): object 'x' not found