Randomly subsample a dataset's mutations.

This functions subsample mutations with uniform probability, retaining all the copy number calls. If data contains driver mutation annotations, these can be forced to remain.

subsample(x, N = 15000, keep_drivers = TRUE)

Arguments

x
An object of class cnaqc, created by the init function.

N
The maximum number of mutations to retain.

keep_drivers
If TRUE, it retains drivers annotated in the data.

Examples

data('example_dataset_CNAqc', package = 'CNAqc')
x = init(example_dataset_CNAqc$snvs, example_dataset_CNAqc$cna,example_dataset_CNAqc$purity)

#>  [ CNAqc - CNA Quality Check ] 
#> 

#> ℹ Using reference genome coordinates for: GRCh38.

#> ℹ Drivers are annotated, but 'gene' column is missing, using mutation location.

#> ! Missing CCF column from CNA calls, adding CCF = 1 assuming clonal CNA calls.

#> ℹ Input n = 12963 mutations for 267 CNA segments (267 clonal, 0 subclonal)

#> 

#> ✔ Mapped n = 12963 mutations to clonal segments (100% of input)

subsample(x, N = 100)

#>  [ CNAqc - CNA Quality Check ] 
#> 

#> ℹ Using reference genome coordinates for: GRCh38.

#> ! Missing CCF column from CNA calls, adding CCF = 1 assuming clonal CNA calls.

#> ℹ Input n = 103 mutations for 267 CNA segments (267 clonal, 0 subclonal)

#> Warning: [CNAqc] a karyotype column is present in CNA calls, and will be overwritten

#> 

#> ✔ Mapped n = 103 mutations to clonal segments (100% of input)

#> $reference_genome
#> [1] "GRCh38"
#> 
#> $snvs
#> # A tibble: 103 x 17
#>    chr       from       to ref   alt   FILTER    DP    NV   VAF ANNOVAR_FUNCTION
#>    <chr>    <dbl>    <dbl> <chr> <chr> <chr>  <dbl> <dbl> <dbl> <chr>           
#>  1 chr2    1.79e8   1.79e8 C     T     PASS     117    77 0.658 exonic          
#>  2 chr16   6.76e7   6.76e7 C     T     PASS     120    54 0.45  exonic          
#>  3 chr17   7.58e6   7.58e6 G     C     PASS      84    78 0.929 exonic          
#>  4 chr5    1.02e8   1.02e8 C     A     PASS     139    61 0.439 upstream        
#>  5 chr7    2.14e7   2.14e7 C     T     PASS     652   470 0.721 intergenic      
#>  6 chr11   1.24e8   1.24e8 C     T     PASS      99    27 0.273 intergenic      
#>  7 chr13   2.64e7   2.64e7 A     T     PASS     142    59 0.415 intronic        
#>  8 chr3    1.56e8   1.56e8 C     T     PASS     180    32 0.178 intronic        
#>  9 chr1    1.03e8   1.03e8 C     T     PASS     172    63 0.366 intergenic      
#> 10 chr1    7.51e7   7.51e7 C     T     PASS     153    91 0.595 intronic        
#> # … with 93 more rows, and 7 more variables: GENE <chr>, is_driver <lgl>,
#> #   driver_label <chr>, gene <chr>, type <chr>, karyotype <chr>,
#> #   segment_id <chr>
#> 
#> $cna
#> # A tibble: 267 x 10
#>    chr      from      to length covRatio Major minor   CCF segment_id          n
#>    <chr>   <int>   <int>  <int>    <dbl> <dbl> <dbl> <dbl> <chr>           <int>
#>  1 chr1   8.40e5  1.69e6 8.50e5     1.19     3     2     1 chr1:840009:16…    NA
#>  2 chr1   1.69e6  1.82e6 1.25e5     1.26     3     2     1 chr1:1689988:1…    NA
#>  3 chr1   1.82e6  9.80e6 7.98e6     1.19     3     2     1 chr1:1815016:9…    NA
#>  4 chr1   1.05e7  1.21e7 1.60e6     1.19     3     2     1 chr1:10479910:…    NA
#>  5 chr1   1.21e7  1.22e7 7.51e4     1.24     3     2     1 chr1:12079917:…    NA
#>  6 chr1   1.22e7  1.28e7 6.85e5     1.19     3     2     1 chr1:12154981:…    NA
#>  7 chr1   1.38e7  1.78e7 4.01e6     1.19     3     2     1 chr1:13780016:…    NA
#>  8 chr1   1.78e7  2.11e7 3.23e6     1.19     3     2     1 chr1:17849962:…     1
#>  9 chr1   2.11e7  2.16e7 4.80e5     1.26     3     2     1 chr1:21080068:…    NA
#> 10 chr1   2.16e7  2.48e7 3.27e6     1.19     3     2     1 chr1:21559998:…    NA
#> # … with 257 more rows
#> 
#> $n_snvs
#> [1] 103
#> 
#> $n_cna
#> [1] 267
#> 
#> $n_cna_clonal
#> [1] 267
#> 
#> $n_cna_sbclonal
#> [1] 0
#> 
#> $n_karyotype
#> 
#>  2:2  2:1  4:2  3:2  3:0  2:0 16:2 
#>   55   17   13   12    3    2    1 
#> 
#> $purity
#> [1] 0.89
#> 
#> $l_karyotype
#>        2:2        2:1        3:2        4:2        3:0        2:0        3:1 
#> 1482910013  420405440  357172683  331457138  137498772   39372558     527401 
#>       25:2      106:1        5:0       69:2       26:2       16:2       70:2 
#>     517926     225714     174002     172894     162913     130712      80103 
#>       99:1 
#>      75782 
#> 
#> $ploidy
#> [1] 4
#> 
#> $most_prevalent_karyotype
#> [1] "2:2"
#> 
#> $basepairs_by_karyotype
#> # A tibble: 15 x 4
#>    minor Major          n karyotype
#>    <dbl> <dbl>      <int> <chr>    
#>  1     2     2 1482910013 2:2      
#>  2     1     2  420405440 2:1      
#>  3     2     3  357172683 3:2      
#>  4     2     4  331457138 4:2      
#>  5     0     3  137498772 3:0      
#>  6     0     2   39372558 2:0      
#>  7     1     3     527401 3:1      
#>  8     2    25     517926 25:2     
#>  9     1   106     225714 106:1    
#> 10     0     5     174002 5:0      
#> 11     2    69     172894 69:2     
#> 12     2    26     162913 26:2     
#> 13     2    16     130712 16:2     
#> 14     2    70      80103 70:2     
#> 15     1    99      75782 99:1     
#> 
#> $most_mutations_karyotype
#> [1] "2:2"
#> 
#> attr(,"class")
#> [1] "cnaqc"

subsample(x, N = 1000)

#>  [ CNAqc - CNA Quality Check ] 
#> 

#> ℹ Using reference genome coordinates for: GRCh38.

#> ! Missing CCF column from CNA calls, adding CCF = 1 assuming clonal CNA calls.

#> ℹ Input n = 1003 mutations for 267 CNA segments (267 clonal, 0 subclonal)

#> Warning: [CNAqc] a karyotype column is present in CNA calls, and will be overwritten

#> 

#> ✔ Mapped n = 1003 mutations to clonal segments (100% of input)

#> $reference_genome
#> [1] "GRCh38"
#> 
#> $snvs
#> # A tibble: 1,003 x 17
#>    chr       from       to ref   alt   FILTER    DP    NV   VAF ANNOVAR_FUNCTION
#>    <chr>    <dbl>    <dbl> <chr> <chr> <chr>  <dbl> <dbl> <dbl> <chr>           
#>  1 chr2    1.79e8   1.79e8 C     T     PASS     117    77 0.658 exonic          
#>  2 chr16   6.76e7   6.76e7 C     T     PASS     120    54 0.45  exonic          
#>  3 chr17   7.58e6   7.58e6 G     C     PASS      84    78 0.929 exonic          
#>  4 chr3    1.91e8   1.91e8 G     T     PASS     196    67 0.342 intergenic      
#>  5 chrX    9.34e7   9.34e7 G     A     PASS     152    35 0.230 intergenic      
#>  6 chr1    8.91e7   8.91e7 G     A     PASS     188   105 0.559 ncRNA_intronic  
#>  7 chr1    1.74e8   1.74e8 T     C     PASS     162    89 0.549 intronic        
#>  8 chrX    1.26e8   1.26e8 C     G     PASS     148    69 0.466 intergenic      
#>  9 chr3    7.85e7   7.85e7 G     A     PASS     211    63 0.299 intergenic      
#> 10 chr14   3.87e7   3.87e7 C     A     PASS     162    69 0.426 intergenic      
#> # … with 993 more rows, and 7 more variables: GENE <chr>, is_driver <lgl>,
#> #   driver_label <chr>, gene <chr>, type <chr>, karyotype <chr>,
#> #   segment_id <chr>
#> 
#> $cna
#> # A tibble: 267 x 10
#>    chr      from      to length covRatio Major minor   CCF segment_id          n
#>    <chr>   <int>   <int>  <int>    <dbl> <dbl> <dbl> <dbl> <chr>           <int>
#>  1 chr1   8.40e5  1.69e6 8.50e5     1.19     3     2     1 chr1:840009:16…    NA
#>  2 chr1   1.69e6  1.82e6 1.25e5     1.26     3     2     1 chr1:1689988:1…    NA
#>  3 chr1   1.82e6  9.80e6 7.98e6     1.19     3     2     1 chr1:1815016:9…     2
#>  4 chr1   1.05e7  1.21e7 1.60e6     1.19     3     2     1 chr1:10479910:…    NA
#>  5 chr1   1.21e7  1.22e7 7.51e4     1.24     3     2     1 chr1:12079917:…    NA
#>  6 chr1   1.22e7  1.28e7 6.85e5     1.19     3     2     1 chr1:12154981:…    NA
#>  7 chr1   1.38e7  1.78e7 4.01e6     1.19     3     2     1 chr1:13780016:…     1
#>  8 chr1   1.78e7  2.11e7 3.23e6     1.19     3     2     1 chr1:17849962:…     2
#>  9 chr1   2.11e7  2.16e7 4.80e5     1.26     3     2     1 chr1:21080068:…    NA
#> 10 chr1   2.16e7  2.48e7 3.27e6     1.19     3     2     1 chr1:21559998:…    NA
#> # … with 257 more rows
#> 
#> $n_snvs
#> [1] 1003
#> 
#> $n_cna
#> [1] 267
#> 
#> $n_cna_clonal
#> [1] 267
#> 
#> $n_cna_sbclonal
#> [1] 0
#> 
#> $n_karyotype
#> 
#>  2:2  4:2  2:1  3:2  3:0  2:0 16:2 
#>  570  148  131  118   28    7    1 
#> 
#> $purity
#> [1] 0.89
#> 
#> $l_karyotype
#>        2:2        2:1        3:2        4:2        3:0        2:0        3:1 
#> 1482910013  420405440  357172683  331457138  137498772   39372558     527401 
#>       25:2      106:1        5:0       69:2       26:2       16:2       70:2 
#>     517926     225714     174002     172894     162913     130712      80103 
#>       99:1 
#>      75782 
#> 
#> $ploidy
#> [1] 4
#> 
#> $most_prevalent_karyotype
#> [1] "2:2"
#> 
#> $basepairs_by_karyotype
#> # A tibble: 15 x 4
#>    minor Major          n karyotype
#>    <dbl> <dbl>      <int> <chr>    
#>  1     2     2 1482910013 2:2      
#>  2     1     2  420405440 2:1      
#>  3     2     3  357172683 3:2      
#>  4     2     4  331457138 4:2      
#>  5     0     3  137498772 3:0      
#>  6     0     2   39372558 2:0      
#>  7     1     3     527401 3:1      
#>  8     2    25     517926 25:2     
#>  9     1   106     225714 106:1    
#> 10     0     5     174002 5:0      
#> 11     2    69     172894 69:2     
#> 12     2    26     162913 26:2     
#> 13     2    16     130712 16:2     
#> 14     2    70      80103 70:2     
#> 15     1    99      75782 99:1     
#> 
#> $most_mutations_karyotype
#> [1] "2:2"
#> 
#> attr(,"class")
#> [1] "cnaqc"