Subset calls by segments' karyotype.

For an object already created, it subsets the calls to those that involve segments with a certain karyotype.

subset_by_segment_karyotype(x, karyotypes)

Arguments

x
An object of class cnaqc, created by the init function.

karyotype
A list of karyotype ids in "Major:minor" notation (e.g., "1:1", "2,1", ...) that will be retained.

Value

An object of class cnaqc, created by the init function.

Examples

data('example_dataset_CNAqc', package = 'CNAqc')
x = init(example_dataset_CNAqc$snvs, example_dataset_CNAqc$cna,example_dataset_CNAqc$purity)

#>  [ CNAqc - CNA Quality Check ] 
#> 

#> ℹ Using reference genome coordinates for: GRCh38.

#> ℹ Drivers are annotated, but 'gene' column is missing, using mutation location.

#> ! Missing CCF column from CNA calls, adding CCF = 1 assuming clonal CNA calls.

#> ℹ Input n = 12963 mutations for 267 CNA segments (267 clonal, 0 subclonal)

#> 

#> ✔ Mapped n = 12963 mutations to clonal segments (100% of input)

subset_by_segment_karyotype(x, '2:2')

#>  [ CNAqc - CNA Quality Check ] 
#> 

#> ℹ Using reference genome coordinates for: GRCh38.

#> ! Missing CCF column from CNA calls, adding CCF = 1 assuming clonal CNA calls.

#> ℹ Input n = 12963 mutations for 80 CNA segments (80 clonal, 0 subclonal)

#> Warning: [CNAqc] a karyotype column is present in CNA calls, and will be overwritten

#> ✖ n = 5485 mutations  cannot be mapped to segments and will be removed.

#> 

#> ✔ Mapped n = 7478 mutations to clonal segments (100% of input)

#> $reference_genome
#> [1] "GRCh38"
#> 
#> $snvs
#> # A tibble: 7,478 x 17
#>    chr      from      to ref   alt   FILTER    DP    NV    VAF ANNOVAR_FUNCTION
#>    <chr>   <dbl>   <dbl> <chr> <chr> <chr>  <dbl> <dbl>  <dbl> <chr>           
#>  1 chr4    54939   54940 C     A     PASS     718    44 0.0613 intronic        
#>  2 chr4   573969  573970 G     C     PASS     164    16 0.0976 intergenic      
#>  3 chr4  1108141 1108142 G     A     PASS     157    66 0.420  ncRNA_intronic  
#>  4 chr4  1395352 1395353 C     T     PASS     178    93 0.522  intergenic      
#>  5 chr4  1537972 1537973 G     A     PASS     155    68 0.439  intergenic      
#>  6 chr4  1565793 1565794 C     T     PASS     172    77 0.448  intergenic      
#>  7 chr4  1600312 1600313 G     A     PASS     139    61 0.439  intergenic      
#>  8 chr4  1801455 1801456 G     A     PASS     161    79 0.491  intronic        
#>  9 chr4  2289980 2289981 C     T     PASS     168    74 0.440  intronic        
#> 10 chr4  2358143 2358144 T     G     PASS     100     6 0.06   intronic        
#> # … with 7,468 more rows, and 7 more variables: GENE <chr>, is_driver <lgl>,
#> #   driver_label <chr>, gene <chr>, type <chr>, karyotype <chr>,
#> #   segment_id <chr>
#> 
#> $cna
#> # A tibble: 80 x 11
#>    chr      from     to length covRatio Major minor k       CCF segment_id     n
#>    <chr>   <int>  <int>  <int>    <dbl> <dbl> <dbl> <chr> <dbl> <chr>      <int>
#>  1 chr1   1.45e8 1.46e8 8.20e5    1.04      2     2 2:2       1 chr1:1449…    NA
#>  2 chr12  1.60e5 3.48e7 3.47e7    0.950     2     2 2:2       1 chr12:160…   131
#>  3 chr12  3.80e7 3.83e7 2.70e5    1.04      2     2 2:2       1 chr12:380…    NA
#>  4 chr12  3.83e7 5.18e7 1.35e7    0.950     2     2 2:2       1 chr12:382…    36
#>  5 chr12  5.20e7 5.21e7 6.44e4    0.950     2     2 2:2       1 chr12:519…    NA
#>  6 chr12  6.74e7 1.33e8 6.61e7    0.950     2     2 2:2       1 chr12:673…   335
#>  7 chr12  1.33e8 1.34e8 3.35e5    1.04      2     2 2:2       1 chr12:133…    NA
#>  8 chr13  1.95e7 1.15e8 9.56e7    0.950     2     2 2:2       1 chr13:194…   479
#>  9 chr14  2.03e7 1.07e8 8.69e7    0.950     2     2 2:2       1 chr14:203…   361
#> 10 chr14  1.07e8 1.07e8 5.33e4    1.04      2     2 2:2       1 chr14:107…    NA
#> # … with 70 more rows
#> 
#> $n_snvs
#> [1] 7478
#> 
#> $n_cna
#> [1] 80
#> 
#> $n_cna_clonal
#> [1] 80
#> 
#> $n_cna_sbclonal
#> [1] 0
#> 
#> $n_karyotype
#>  2:2 
#> 7478 
#> 
#> $purity
#> [1] 0.89
#> 
#> $l_karyotype
#>        2:2 
#> 1482910013 
#> 
#> $ploidy
#> [1] 4
#> 
#> $most_prevalent_karyotype
#> [1] "2:2"
#> 
#> $basepairs_by_karyotype
#> # A tibble: 1 x 4
#>   minor Major          n karyotype
#>   <dbl> <dbl>      <int> <chr>    
#> 1     2     2 1482910013 2:2      
#> 
#> $most_mutations_karyotype
#> [1] "2:2"
#> 
#> attr(,"class")
#> [1] "cnaqc"

subset_by_segment_karyotype(x, '2:1')

#>  [ CNAqc - CNA Quality Check ] 
#> 

#> ℹ Using reference genome coordinates for: GRCh38.

#> ! Missing CCF column from CNA calls, adding CCF = 1 assuming clonal CNA calls.

#> ℹ Input n = 12963 mutations for 49 CNA segments (49 clonal, 0 subclonal)

#> Warning: [CNAqc] a karyotype column is present in CNA calls, and will be overwritten

#> ✖ n = 11400 mutations  cannot be mapped to segments and will be removed.

#> 

#> ✔ Mapped n = 1563 mutations to clonal segments (100% of input)

#> $reference_genome
#> [1] "GRCh38"
#> 
#> $snvs
#> # A tibble: 1,563 x 17
#>    chr      from      to ref   alt   FILTER    DP    NV    VAF ANNOVAR_FUNCTION
#>    <chr>   <dbl>   <dbl> <chr> <chr> <chr>  <dbl> <dbl>  <dbl> <chr>           
#>  1 chr2   357969  357970 C     A     PASS     104    58 0.558  intergenic      
#>  2 chr2   909304  909305 A     G     PASS      15     7 0.467  ncRNA_intronic  
#>  3 chr2  1035751 1035752 C     T     PASS      93    57 0.613  intronic        
#>  4 chr2  1326719 1326720 A     T     PASS     104    55 0.529  intronic        
#>  5 chr2  1515962 1515963 C     T     PASS      90    50 0.556  intronic        
#>  6 chr2  2198361 2198362 G     T     PASS     125    42 0.336  intronic        
#>  7 chr2  2898536 2898537 C     T     PASS     109    58 0.532  downstream      
#>  8 chr2  3125481 3125482 A     G     PASS     134    35 0.261  ncRNA_intronic  
#>  9 chr2  3832360 3832361 A     T     PASS     120    68 0.567  intergenic      
#> 10 chr2  3878408 3878409 T     A     PASS     126    10 0.0794 intergenic      
#> # … with 1,553 more rows, and 7 more variables: GENE <chr>, is_driver <lgl>,
#> #   driver_label <chr>, gene <chr>, type <chr>, karyotype <chr>,
#> #   segment_id <chr>
#> 
#> $cna
#> # A tibble: 49 x 11
#>    chr      from     to length covRatio Major minor k       CCF segment_id     n
#>    <chr>   <int>  <int>  <int>    <dbl> <dbl> <dbl> <chr> <dbl> <chr>      <int>
#>  1 chr11  1.90e5 2.63e6 2.44e6    0.731     2     1 2:1       1 chr11:190…     9
#>  2 chr11  2.63e6 2.68e6 5.01e4    0.748     2     1 2:1       1 chr11:262…    NA
#>  3 chr11  2.68e6 4.80e6 2.12e6    0.731     2     1 2:1       1 chr11:267…     6
#>  4 chr11  4.91e6 9.94e6 5.03e6    0.731     2     1 2:1       1 chr11:490…    12
#>  5 chr11  9.97e6 1.71e7 7.10e6    0.731     2     1 2:1       1 chr11:997…    12
#>  6 chr11  1.71e7 1.74e7 3.25e5    0.748     2     1 2:1       1 chr11:170…     3
#>  7 chr11  1.74e7 4.89e7 3.15e7    0.731     2     1 2:1       1 chr11:173…   141
#>  8 chr11  5.02e7 5.06e7 4.00e5    0.731     2     1 2:1       1 chr11:501…     2
#>  9 chr11  5.12e7 5.16e7 3.50e5    0.748     2     1 2:1       1 chr11:512…     1
#> 10 chr11  5.50e7 5.54e7 3.60e5    0.731     2     1 2:1       1 chr11:549…     1
#> # … with 39 more rows
#> 
#> $n_snvs
#> [1] 1563
#> 
#> $n_cna
#> [1] 49
#> 
#> $n_cna_clonal
#> [1] 49
#> 
#> $n_cna_sbclonal
#> [1] 0
#> 
#> $n_karyotype
#>  2:1 
#> 1563 
#> 
#> $purity
#> [1] 0.89
#> 
#> $l_karyotype
#>       2:1 
#> 420405440 
#> 
#> $ploidy
#> [1] 3
#> 
#> $most_prevalent_karyotype
#> [1] "2:1"
#> 
#> $basepairs_by_karyotype
#> # A tibble: 1 x 4
#>   minor Major         n karyotype
#>   <dbl> <dbl>     <int> <chr>    
#> 1     1     2 420405440 2:1      
#> 
#> $most_mutations_karyotype
#> [1] "2:1"
#> 
#> attr(,"class")
#> [1] "cnaqc"