Retain clonal segments with minimum number of mutations.

It retains only the set of clonal segments that have mapped at least `n` mutations mapped.

subset_by_segment_minmutations(x, n = 50)

Arguments

x

A CNAqc object.

n

The minimum number of mutations per segment.

Value

A CNAqc object that should have less segments.

Examples

data('example_dataset_CNAqc', package = 'CNAqc')
x = init(mutations = example_dataset_CNAqc$mutations, cna = example_dataset_CNAqc$cna, purity = example_dataset_CNAqc$purity)
#> 
#> ── CNAqc - CNA Quality Check ───────────────────────────────────────────────────
#> 
#> ℹ Using reference genome coordinates for: GRCh38.
#> ✔ Found annotated driver mutations: TTN, CTCF, and TP53.
#> ✔ Fortified calls for 12963 somatic mutations: 12963 SNVs (100%) and 0 indels.
#> ! CNAs have no CCF, assuming clonal CNAs (CCF = 1).
#> ✔ Fortified CNAs for 267 segments: 267 clonal and 0 subclonal.
#> ✔ 12963 mutations mapped to clonal CNAs.

subset_by_segment_minmutations(x, 0)
#> ! This filter has no effect on the input data.
#> ── [ CNAqc ] MySample 12963 mutations in 267 segments (267 clonal, 0 subclonal).
#> 
#> ── Clonal CNAs 
#> 
#>    2:2  [n = 7478, L = 1483 Mb] ■■■■■■■■■■■■■■■■■■■■■■■■■■■  { CTCF }
#>    4:2  [n = 1893, L =  331 Mb] ■■■■■■■
#>    3:2  [n = 1625, L =  357 Mb] ■■■■■■
#>    2:1  [n = 1563, L =  420 Mb] ■■■■■■  { TTN }
#>    3:0  [n =  312, L =  137 Mb] ■
#>    2:0  [n =   81, L =   39 Mb]   { TP53 }
#>   16:2  [n =    4, L =    0 Mb] 
#>   25:2  [n =    2, L =    1 Mb] 
#>    3:1  [n =    2, L =    1 Mb] 
#>  106:1  [n =    1, L =    0 Mb] 
#> 
#> ℹ Sample Purity: 89% ~ Ploidy: 4.
#> 
#> ℹ There are 3 annotated driver(s) mapped to clonal CNAs.
#>          chr      from        to ref alt  DP NV       VAF driver_label is_driver
#>         chr2 179431633 179431634   C   T 117 77 0.6581197          TTN      TRUE
#>        chr16  67646006  67646007   C   T 120 54 0.4500000         CTCF      TRUE
#>        chr17   7577106   7577107   G   C  84 78 0.9285714         TP53      TRUE
subset_by_segment_minmutations(x, 10)
#> ℹ Dropping 79 clonal segments with less than 10 mutations - losing 272/12963 mutations (2.0982797192008%)
#> 
#> ── CNAqc - CNA Quality Check ───────────────────────────────────────────────────
#> 
#> ℹ Using reference genome coordinates for: GRCh38.
#> ✔ Found annotated driver mutations: TTN, CTCF, and TP53.
#> ✔ Fortified calls for 12963 somatic mutations: 12963 SNVs (100%) and 0 indels.
#> ✔ Fortified CNAs for 147 segments: 147 clonal and 0 subclonal.
#> Warning: [CNAqc] a karyotype column is present in CNA calls, and will be overwritten
#> ✔ 12691 mutations mapped to clonal CNAs.
#> 
#>               ┌ Driver(s):  TP53 ─────────────────────────────────┐
#>               │                                                   │
#>               │   Driver cannot be mapped - out of any segment!   │
#>               │                                                   │
#>               └───────────────────────────────────────────────────┘
#> ── [ CNAqc ] MySample 12691 mutations in 147 segments (147 clonal, 0 subclonal).
#> 
#> ── Clonal CNAs 
#> 
#>  2:2  [n = 7399, L = 1452 Mb] ■■■■■■■■■■■■■■■■■■■■■■■■■■■  { CTCF }
#>  4:2  [n = 1874, L =  327 Mb] ■■■■■■■
#>  3:2  [n = 1537, L =  321 Mb] ■■■■■■
#>  2:1  [n = 1510, L =  400 Mb] ■■■■■  { TTN }
#>  3:0  [n =  311, L =  136 Mb] ■
#>  2:0  [n =   60, L =   26 Mb] 
#> 
#> ℹ Sample Purity: 89% ~ Ploidy: 4.
#> 
#> ℹ There are 2 annotated driver(s) mapped to clonal CNAs.
#>          chr      from        to ref alt  DP NV       VAF driver_label is_driver
#>         chr2 179431633 179431634   C   T 117 77 0.6581197          TTN      TRUE
#>        chr16  67646006  67646007   C   T 120 54 0.4500000         CTCF      TRUE