Extract CCF estimates.

This function extracts Cancer Cell Fractions (CCFs) estimates from a CNAqc object, after they have been computed using function `compute_CCF`. The estimates are pooled across the used karyotypes, and extracted from the `CCF_estimates` field of the input object.

CCF(x)

Arguments

x

A CNAqc object.

Value

A mutations tibble with columns for mutation multiplicity and CCFs.

Examples

data("example_PCAWG", package = 'CNAqc')
CCF(example_PCAWG)
#> # A tibble: 120,065 × 60
#>    chr        from        to ref   alt   sample       icgc_donor_id project_code
#>    <chr>     <dbl>     <dbl> <chr> <chr> <chr>        <chr>         <chr>       
#>  1 chr1  240197260 240197260 G     A     51893d3f-e7… DO220903      MELA-AU     
#>  2 chr1  240197743 240197743 T     -     51893d3f-e7… DO220903      MELA-AU     
#>  3 chr1  240209355 240209355 G     A     51893d3f-e7… DO220903      MELA-AU     
#>  4 chr10  42871855  42871855 G     A     51893d3f-e7… DO220903      MELA-AU     
#>  5 chr10  42874161  42874161 G     A     51893d3f-e7… DO220903      MELA-AU     
#>  6 chr10  42886327  42886327 C     T     51893d3f-e7… DO220903      MELA-AU     
#>  7 chr10  42887738  42887738 G     A     51893d3f-e7… DO220903      MELA-AU     
#>  8 chr10  42888438  42888438 C     T     51893d3f-e7… DO220903      MELA-AU     
#>  9 chr10  42888486  42888486 G     A     51893d3f-e7… DO220903      MELA-AU     
#> 10 chr10  42891609  42891609 C     T     51893d3f-e7… DO220903      MELA-AU     
#> # ℹ 120,055 more rows
#> # ℹ 52 more variables: icgc_specimen_id <chr>, icgc_sample_id <chr>,
#> #   submitted_sample_id <chr>, submitted_matched_sample_id <chr>,
#> #   chromosome_strand <dbl>, assembly_version <chr>, mutation_type <chr>,
#> #   reference_genome_allele <chr>, quality_score <lgl>, probability <lgl>,
#> #   verification_status <chr>, verification_platform <lgl>,
#> #   biological_validation_status <chr>, biological_validation_platform <lgl>, …