Return the available mutation CCF estimates in the data.

This function extracts Cancer Cell Fractions (CCFs) estimates from a `CNAqc` object, after they have been computed using function `compute_CCF`. The estimates are pooled across the used karyotypes, and extracted from the `CCF_estimates` field of the input object.

CCF(x)

Arguments

x

An object of class cnaqc, where CCF have been computed using function `compute_CCF`.

Value

A tibble with the mutations with the new columns for mutation multiplicity and CCF values.

Examples

data('example_dataset_CNAqc')
x = init(example_dataset_CNAqc$snvs, example_dataset_CNAqc$cna, example_dataset_CNAqc$purity)
#>  [ CNAqc - CNA Quality Check ] 
#> 
#> ℹ Using reference genome coordinates for: GRCh38.
#> ℹ Drivers are annotated, but 'gene' column is missing, using mutation location.
#> ✔ Fortified calls for 12963 somatic mutations: 12963 SNVs (100%) and 0 indels.
#> ! CNAs have no CCF, assuming clonal CNAs (CCF = 1).
#> ✔ Fortified CNAs for 267 segments: 267 clonal and 0 subclonal.
#> ✔ 12963 mutations mapped to clonal CNAs.

x = compute_CCF(x)
#> Warning: Some karyotypes have fewer than25and will not be analysed.
#> ── Computing mutation multiplicity for karyotype 2:0 using the entropy method. ─
#> ℹ Expected Binomial peak(s) for these calls (1 and 2 copies): 0.445 and 0.89
#> ℹ Mixing pre/ post aneuploidy: 0.09 and 0.91
#> ℹ Not assignamble area: [0.631578947368421; 0.723684210526316]
#> ── Computing mutation multiplicity for karyotype 2:1 using the entropy method. ─
#> ℹ Expected Binomial peak(s) for these calls (1 and 2 copies): 0.307958477508651 and 0.615916955017301
#> ℹ Mixing pre/ post aneuploidy: 0.55 and 0.45
#> ℹ Not assignamble area: [0.423423423423423; 0.504504504504504]
#> ── Computing mutation multiplicity for karyotype 2:2 using the entropy method. ─
#> ℹ Expected Binomial peak(s) for these calls (1 and 2 copies): 0.235449735449735 and 0.470899470899471
#> ℹ Mixing pre/ post aneuploidy: 0.09 and 0.91
#> ℹ Not assignamble area: [0.290780141843972; 0.368794326241135]
CCF(x)
#> # A tibble: 9,122 × 19
#>    chr       from       to ref   alt   FILTER    DP    NV   VAF ANNOVAR_…¹ GENE 
#>    <chr>    <dbl>    <dbl> <chr> <chr> <chr>  <dbl> <dbl> <dbl> <chr>      <chr>
#>  1 chr12 52097806 52097807 A     C     PASS      75     6 0.08  intronic   SCN8A
#>  2 chr12 52518721 52518722 G     A     PASS      60    53 0.883 intergenic AC07…
#>  3 chr12 53360455 53360456 C     T     PASS      82    75 0.915 intergenic KRT1…
#>  4 chr12 54790107 54790108 C     T     PASS      86    77 0.895 exonic     ITGA5
#>  5 chr12 54790131 54790132 A     T     PASS      72    62 0.861 exonic     ITGA5
#>  6 chr12 55400463 55400464 T     A     PASS      85    40 0.471 intergenic TESP…
#>  7 chr12 57769765 57769766 C     T     PASS      73    62 0.849 intronic   R3HD…
#>  8 chr12 58555290 58555291 C     T     PASS      65    51 0.785 intergenic RP11…
#>  9 chr12 58579121 58579122 T     A     PASS      64    57 0.891 intergenic RP11…
#> 10 chr12 59011655 59011656 G     A     PASS      54    44 0.815 ncRNA_int… RP11…
#> # … with 9,112 more rows, 8 more variables: is_driver <lgl>,
#> #   driver_label <chr>, gene <chr>, type <chr>, karyotype <chr>,
#> #   segment_id <chr>, mutation_multiplicity <dbl>, CCF <dbl>, and abbreviated
#> #   variable name ¹​ANNOVAR_FUNCTION