Analyze calls by peak detection.

This function carries out a peak-detection analysis based on a joint criterion: KDE (package peakPick) plus Binomial mixture (package BMix). This is used to determine if the clonal mutations - for a given karyotype - have VAF close to the expected clonal VAF, computed from input major and minor alleles and tumour purity. A score is produced as a linear combination of the distance of the actual peak to the expected one.

analyze_peaks(
  x,
  karyotypes = c("1:0", "1:1", "2:0", "2:1", "2:2"),
  min_karyotype_size = 0,
  min_absolute_karyotype_mutations = 10,
  p_binsize_peaks = 0.005,
  matching_epsilon = NULL,
  purity_error = 0.05,
  VAF_tolerance = 0.015,
  n_bootstrap = 1,
  kernel_adjust = 1,
  matching_strategy = "closest",
  KDE = TRUE
)

Arguments

x
An object of class cnaqc, created by the init function.

karyotypes
The list of karyotypes to test. By default LOH regions (A, AA), diploid regions (AB), and amplification regions (AAB, AABB). These correspond to '1:0', '1:1', '2:1', '2:0', '2:2' in "Major:minor" notation.

min_karyotype_size
Karyotypes are subset by their size (normalized for the number of input mutations). Karyotypes smaller than this value are removed from analysis.

min_absolute_karyotype_mutations
As min_karyotype_size this imposes a minimum number of mutations on a karyotype to be analysed (in absolute number of mutations). Karyotypes smaller than this value are removed from analysis.

p_binsize_peaks
Peaks detected will be filtered if, in a peak, we map less than p_binsize_peaks * N mutations. The value N is obtained couting all mutations that map in all peaks.

matching_epsilon
Peaks at a location are matched with a range of epsilon * 0.5, left and right. By default (0.025) a maxium 5% tolerance is adopted. The status of overall PASS/ FAIL for this analysis is determined with the same tolerance, given the score computed across all karyotypes.

n_bootstrap
Number of times peak detection is bootstrapped (this helps sometimes finding peaks that might be visually observable but fail to be detected by the underlying peak detection heuristic). The default of this parameter is 1, meaning that no bootstrap is performed.

kernel_adjust
KDE adjust density parameter; see density. A Gaussian kernel is used (kernel = 'gaussian').

matching_strategy
If "closest", the closest peak will be used to match the expected peak. If "rightmost", peaks are matched prioritizing right to left peaks (the higher get matched first); this strategy is more correct in principle but works only if there are no spurious peaks in the estimated density. By defaule the first strategy is used.

Value

An object of class cnaqc, modified to hold the results from this analysis. See the vignette to see how to extract and plot the results.

Examples

data('example_dataset_CNAqc', package = 'CNAqc')
x = init(example_dataset_CNAqc$snvs, example_dataset_CNAqc$cna,example_dataset_CNAqc$purity)
#>  [ CNAqc - CNA Quality Check ] 
#> 
#> ℹ Using reference genome coordinates for: GRCh38.
#> ℹ Drivers are annotated, but 'gene' column is missing, using mutation location.
#> ! Missing CCF column from CNA calls, adding CCF = 1 assuming clonal CNA calls.
#> ℹ Input n = 12963 mutations for 267 CNA segments (267 clonal, 0 subclonal)
#> 
#> ✓ Mapped n = 12963 mutations to clonal segments (100% of input)

x = analyze_peaks(x)
#> ℹ Karyotypes 1:0, 1:1, 2:0, 2:1, and 2:2. Matching strategy closest. KDE = TRUE.
#> ℹ Found n = 9122 mutations in 2:2, 2:1, 2:0 (skipping those with n < 0 mutations).
#> ! Some peaks have been removed
#> # A tibble: 2 x 5
#>       x     y counts_per_bin discarded from 
#>   <dbl> <dbl>          <int> <lgl>     <chr>
#> 1  0.13   0.3              1 TRUE      KDE  
#> 2  0.32   0.3              0 TRUE      KDE  
#> ! Some peaks have been removed
#> # A tibble: 10 x 5
#>        x     y counts_per_bin discarded from 
#>    <dbl> <dbl>          <int> <lgl>     <chr>
#>  1  0.11  0.27             20 TRUE      KDE  
#>  2  0.67  0.01              0 TRUE      KDE  
#>  3  0.72  0.02              2 TRUE      KDE  
#>  4  0.75  0.01              2 TRUE      KDE  
#>  5  0.79  0.01              1 TRUE      KDE  
#>  6  0.82  0.01              1 TRUE      KDE  
#>  7  0.85  0.01              1 TRUE      KDE  
#>  8  0.91  0.01              1 TRUE      KDE  
#>  9  0.93  0.01              0 TRUE      KDE  
#> 10  0.97  0.02              0 TRUE      KDE  
#> # A tibble: 6 x 17
#>   mutation_multiplic… karyotype  peak     x     y counts_per_bin discarded from 
#>                 <dbl> <chr>     <dbl> <dbl> <dbl>          <int> <lgl>     <chr>
#> 1                   1 2:0       0.445 0.423 0.172              0 FALSE     BMix 
#> 2                   2 2:0       0.89  0.885 7.99               7 FALSE     BMix 
#> 3                   1 2:1       0.308 0.307 3.60              59 FALSE     BMix 
#> 4                   2 2:1       0.616 0.602 2.94              60 FALSE     BMix 
#> 5                   1 2:2       0.235 0.24  0.86              65 FALSE     KDE  
#> 6                   2 2:2       0.471 0.463 8.11             625 FALSE     BMix 
#> # … with 9 more variables: offset_VAF <dbl>, offset <dbl>, weight <dbl>,
#> #   epsilon <dbl>, VAF_tolerance <dbl>, matched <lgl>, score <dbl>,
#> #   purity_error <dbl>, QC <chr>
#> ✓ Peak detection PASS with r = 0.000739884902244746 and tolerance e = 
print(x)
#> ── [ CNAqc ]  n = 12963 mutations in 267 segments (267 clonal + 0 subclonal). Ge
#> 
#>    2:2  [n = 7478, L = 1483 Mb] ■■■■■■■■■■■■■■■■■■■■■■■■■■■  { CTCF }
#>    4:2  [n = 1893, L =  331 Mb] ■■■■■■■
#>    3:2  [n = 1625, L =  357 Mb] ■■■■■■
#>    2:1  [n = 1563, L =  420 Mb] ■■■■■■  { TTN }
#>    3:0  [n =  312, L =  137 Mb] ■
#>    2:0  [n =   81, L =   39 Mb]   { TP53 }
#>   16:2  [n =    4, L =    0 Mb] 
#>   25:2  [n =    2, L =    1 Mb] 
#>    3:1  [n =    2, L =    1 Mb] 
#>  106:1  [n =    1, L =    0 Mb] 
#>   26:2  [n =    1, L =    0 Mb] 
#>   99:1  [n =    1, L =    0 Mb] 
#> 
#> ℹ Sample Purity: 89% ~ Ploidy: 4.
#> 
#> ℹ There are 3 annotated driver(s).
#>          chr      from        to ref alt  DP NV       VAF driver_label is_driver
#>         chr2 179431633 179431634   C   T 117 77 0.6581197          TTN      TRUE
#>        chr16  67646006  67646007   C   T 120 54 0.4500000         CTCF      TRUE
#>        chr17   7577106   7577107   G   C  84 78 0.9285714         TP53      TRUE
#> 
#> ──  PASS  Peaks QC closest: 100%, λ = 7e-04. Purity correction: 0%. 
#> ℹ 2:0 ~ n = 81    ( 1%) →  PASS  0.043    PASS  0.005  
#> ℹ 2:1 ~ n = 1563  (17%) →  PASS  0.003    PASS  0.028  
#> ℹ 2:2 ~ n = 7478  (82%) →  PASS  -0.034   PASS  0.028  

print(x$peaks_analysis)
#> $score
#> [1] 0.0007398849
#> 
#> $fits
#> $fits$`1:0`
#> $fits$`1:0`$matching
#> NULL
#> 
#> 
#> $fits$`1:1`
#> $fits$`1:1`$matching
#> NULL
#> 
#> 
#> $fits$`2:0`
#> $fits$`2:0`$matching
#> # A tibble: 2 x 14
#> # Rowwise: 
#>   mutation_multiplic… karyotype  peak     x     y counts_per_bin discarded from 
#>                 <dbl> <chr>     <dbl> <dbl> <dbl>          <int> <lgl>     <chr>
#> 1                   1 2:0       0.445 0.423 0.172              0 FALSE     BMix 
#> 2                   2 2:0       0.89  0.885 7.99               7 FALSE     BMix 
#> # … with 6 more variables: offset_VAF <dbl>, offset <dbl>, weight <dbl>,
#> #   epsilon <dbl>, VAF_tolerance <dbl>, matched <lgl>
#> 
#> $fits$`2:0`$density
#> 
#> Call:
#> 	density.default(x = y, adjust = kernel_adjust, kernel = "gaussian")
#> 
#> Data: y (81 obs.);	Bandwidth 'bw' = 0.0165
#> 
#>        x                y           
#>  Min.   :0.0305   Min.   :0.000000  
#>  1st Qu.:0.2815   1st Qu.:0.000732  
#>  Median :0.5325   Median :0.193553  
#>  Mean   :0.5325   Mean   :0.994921  
#>  3rd Qu.:0.7836   3rd Qu.:0.556356  
#>  Max.   :1.0346   Max.   :9.001287  
#> 
#> $fits$`2:0`$xy_peaks
#> # A tibble: 8 x 5
#>       x     y counts_per_bin discarded from 
#>   <dbl> <dbl>          <int> <lgl>     <chr>
#> 1 0.13  0.3                1 TRUE      KDE  
#> 2 0.32  0.3                0 TRUE      KDE  
#> 3 0.38  0.56               1 FALSE     KDE  
#> 4 0.47  0.79               1 FALSE     KDE  
#> 5 0.87  9                  5 FALSE     KDE  
#> 6 0.101 0.199              0 FALSE     BMix 
#> 7 0.423 0.172              0 FALSE     BMix 
#> 8 0.885 7.99               7 FALSE     BMix 
#> 
#> 
#> $fits$`2:1`
#> $fits$`2:1`$matching
#> # A tibble: 2 x 14
#> # Rowwise: 
#>   mutation_multiplic… karyotype  peak     x     y counts_per_bin discarded from 
#>                 <dbl> <chr>     <dbl> <dbl> <dbl>          <int> <lgl>     <chr>
#> 1                   1 2:1       0.308 0.307  3.60             59 FALSE     BMix 
#> 2                   2 2:1       0.616 0.602  2.94             60 FALSE     BMix 
#> # … with 6 more variables: offset_VAF <dbl>, offset <dbl>, weight <dbl>,
#> #   epsilon <dbl>, VAF_tolerance <dbl>, matched <lgl>
#> 
#> $fits$`2:1`$density
#> 
#> Call:
#> 	density.default(x = y, adjust = kernel_adjust, kernel = "gaussian")
#> 
#> Data: y (1563 obs.);	Bandwidth 'bw' = 0.0341
#> 
#>        x                  y          
#>  Min.   :-0.05578   Min.   :0.00016  
#>  1st Qu.: 0.17384   1st Qu.:0.21293  
#>  Median : 0.40346   Median :0.51690  
#>  Mean   : 0.40346   Mean   :1.08768  
#>  3rd Qu.: 0.63308   3rd Qu.:1.99965  
#>  Max.   : 0.86270   Max.   :3.60365  
#> 
#> $fits$`2:1`$xy_peaks
#> # A tibble: 6 x 5
#>       x     y counts_per_bin discarded from 
#>   <dbl> <dbl>          <int> <lgl>     <chr>
#> 1 0.08  0.27               5 FALSE     KDE  
#> 2 0.3   3.6               78 FALSE     KDE  
#> 3 0.6   2.94              60 FALSE     KDE  
#> 4 0.602 2.94              60 FALSE     BMix 
#> 5 0.307 3.60              59 FALSE     BMix 
#> 6 0.106 0.248              2 FALSE     BMix 
#> 
#> 
#> $fits$`2:2`
#> $fits$`2:2`$matching
#> # A tibble: 2 x 14
#> # Rowwise: 
#>   mutation_multiplic… karyotype  peak     x     y counts_per_bin discarded from 
#>                 <dbl> <chr>     <dbl> <dbl> <dbl>          <int> <lgl>     <chr>
#> 1                   1 2:2       0.235 0.24   0.86             65 FALSE     KDE  
#> 2                   2 2:2       0.471 0.463  8.11            625 FALSE     BMix 
#> # … with 6 more variables: offset_VAF <dbl>, offset <dbl>, weight <dbl>,
#> #   epsilon <dbl>, VAF_tolerance <dbl>, matched <lgl>
#> 
#> $fits$`2:2`$density
#> 
#> Call:
#> 	density.default(x = y, adjust = kernel_adjust, kernel = "gaussian")
#> 
#> Data: y (7478 obs.);	Bandwidth 'bw' = 0.008033
#> 
#>        x                 y           
#>  Min.   :0.02101   Min.   :0.000244  
#>  1st Qu.:0.27179   1st Qu.:0.007045  
#>  Median :0.52256   Median :0.203133  
#>  Mean   :0.52256   Mean   :0.995946  
#>  3rd Qu.:0.77333   3rd Qu.:0.787272  
#>  Max.   :1.02410   Max.   :8.138867  
#> 
#> $fits$`2:2`$xy_peaks
#> # A tibble: 15 x 5
#>        x     y counts_per_bin discarded from 
#>    <dbl> <dbl>          <int> <lgl>     <chr>
#>  1 0.11  0.27              20 TRUE      KDE  
#>  2 0.22  0.84              69 FALSE     KDE  
#>  3 0.24  0.86              65 FALSE     KDE  
#>  4 0.46  8.14             625 FALSE     KDE  
#>  5 0.67  0.01               0 TRUE      KDE  
#>  6 0.72  0.02               2 TRUE      KDE  
#>  7 0.75  0.01               2 TRUE      KDE  
#>  8 0.79  0.01               1 TRUE      KDE  
#>  9 0.82  0.01               1 TRUE      KDE  
#> 10 0.85  0.01               1 TRUE      KDE  
#> 11 0.91  0.01               1 TRUE      KDE  
#> 12 0.93  0.01               0 TRUE      KDE  
#> 13 0.97  0.02               0 TRUE      KDE  
#> 14 0.463 8.11             625 FALSE     BMix 
#> 15 0.176 0.447             33 FALSE     BMix 
#> 
#> 
#> 
#> $matches
#> # A tibble: 6 x 17
#>   mutation_multiplic… karyotype  peak     x     y counts_per_bin discarded from 
#>                 <dbl> <chr>     <dbl> <dbl> <dbl>          <int> <lgl>     <chr>
#> 1                   1 2:0       0.445 0.423 0.172              0 FALSE     BMix 
#> 2                   2 2:0       0.89  0.885 7.99               7 FALSE     BMix 
#> 3                   1 2:1       0.308 0.307 3.60              59 FALSE     BMix 
#> 4                   2 2:1       0.616 0.602 2.94              60 FALSE     BMix 
#> 5                   1 2:2       0.235 0.24  0.86              65 FALSE     KDE  
#> 6                   2 2:2       0.471 0.463 8.11             625 FALSE     BMix 
#> # … with 9 more variables: offset_VAF <dbl>, offset <dbl>, weight <dbl>,
#> #   epsilon <dbl>, VAF_tolerance <dbl>, matched <lgl>, score <dbl>,
#> #   purity_error <dbl>, QC <chr>
#> 
#> $matching_strategy
#> [1] "closest"
#> 
#> $purity_error
#> [1] 0.05
#> 
#> $min_karyotype_size
#> [1] 0
#> 
#> $p_binsize_peaks
#> [1] 0.005
#> 
#> $QC
#> [1] "PASS"
#> 
#> $KDE
#> [1] TRUE
#> 

x = analyze_peaks(x, matching_strategy = "rightmost")
#> ℹ Karyotypes 1:0, 1:1, 2:0, 2:1, and 2:2. Matching strategy rightmost. KDE = TRUE.
#> ℹ Found n = 9122 mutations in 2:2, 2:1, 2:0 (skipping those with n < 0 mutations).
#> ! Some KDE peaks have been removed
#> # A tibble: 2 x 5
#>       x     y counts_per_bin discarded from 
#>   <dbl> <dbl>          <int> <lgl>     <chr>
#> 1  0.13   0.3              1 TRUE      KDE  
#> 2  0.32   0.3              0 TRUE      KDE  
#> Warning: the condition has length > 1 and only the first element will be used
#> Warning: the condition has length > 1 and only the first element will be used
#> ! Some KDE peaks have been removed
#> # A tibble: 10 x 5
#>        x     y counts_per_bin discarded from 
#>    <dbl> <dbl>          <int> <lgl>     <chr>
#>  1  0.11  0.27             20 TRUE      KDE  
#>  2  0.67  0.01              0 TRUE      KDE  
#>  3  0.72  0.02              2 TRUE      KDE  
#>  4  0.75  0.01              2 TRUE      KDE  
#>  5  0.79  0.01              1 TRUE      KDE  
#>  6  0.82  0.01              1 TRUE      KDE  
#>  7  0.85  0.01              1 TRUE      KDE  
#>  8  0.91  0.01              1 TRUE      KDE  
#>  9  0.93  0.01              0 TRUE      KDE  
#> 10  0.97  0.02              0 TRUE      KDE  
#> Warning: the condition has length > 1 and only the first element will be used
#> # A tibble: 6 x 17
#>   mutation_multiplic… karyotype  peak     x     y counts_per_bin discarded from 
#>                 <dbl> <chr>     <dbl> <dbl> <dbl>          <int> <lgl>     <chr>
#> 1                   2 2:0       0.89  0.885  7.99              7 FALSE     BMix 
#> 2                   1 2:0       0.445 0.87   9                 5 FALSE     KDE  
#> 3                   2 2:1       0.616 0.602  2.94             60 FALSE     BMix 
#> 4                   1 2:1       0.308 0.6    2.94             60 FALSE     KDE  
#> 5                   2 2:2       0.471 0.461  8.14            625 FALSE     BMix 
#> 6                   1 2:2       0.235 0.46   8.14            625 FALSE     KDE  
#> # … with 9 more variables: offset_VAF <dbl>, offset <dbl>, weight <dbl>,
#> #   epsilon <dbl>, VAF_tolerance <dbl>, matched <lgl>, score <dbl>,
#> #   purity_error <dbl>, QC <chr>
#> ✓ Peak detection PASS with r = -29.0623856969481 and tolerance e = 
print(x)
#> ── [ CNAqc ]  n = 12963 mutations in 267 segments (267 clonal + 0 subclonal). Ge
#> 
#>    2:2  [n = 7478, L = 1483 Mb] ■■■■■■■■■■■■■■■■■■■■■■■■■■■  { CTCF }
#>    4:2  [n = 1893, L =  331 Mb] ■■■■■■■
#>    3:2  [n = 1625, L =  357 Mb] ■■■■■■
#>    2:1  [n = 1563, L =  420 Mb] ■■■■■■  { TTN }
#>    3:0  [n =  312, L =  137 Mb] ■
#>    2:0  [n =   81, L =   39 Mb]   { TP53 }
#>   16:2  [n =    4, L =    0 Mb] 
#>   25:2  [n =    2, L =    1 Mb] 
#>    3:1  [n =    2, L =    1 Mb] 
#>  106:1  [n =    1, L =    0 Mb] 
#>   26:2  [n =    1, L =    0 Mb] 
#>   99:1  [n =    1, L =    0 Mb] 
#> 
#> ℹ Sample Purity: 89% ~ Ploidy: 4.
#> 
#> ℹ There are 3 annotated driver(s).
#>          chr      from        to ref alt  DP NV       VAF driver_label is_driver
#>         chr2 179431633 179431634   C   T 117 77 0.6581197          TTN      TRUE
#>        chr16  67646006  67646007   C   T 120 54 0.4500000         CTCF      TRUE
#>        chr17   7577106   7577107   G   C  84 78 0.9285714         TP53      TRUE
#> 
#> ──  PASS  Peaks QC rightmost: 100%, λ = -29.0624. Purity correction: -2906%. 
#> ℹ 2:0 ~ n = 81    ( 1%) →  PASS  0.005    PASS  -0.85  
#> ℹ 2:1 ~ n = 1563  (17%) →  PASS  0.028    PASS  -3.651 
#> ℹ 2:2 ~ n = 7478  (82%) →  PASS  0.035    PASS  -70.172