Analyze CNA calls by peak detection.

This function carries out a peak-detection analysis based on KDE and the package peakPick in order to determine if the mutations that map to a certain karyotype have an allelic frequency that is consistent with the Major and minor alleles of the segment, and the tumour purity. A score is produced as a linear combination of the distance of the actual peak to the expected one, derived with standard equations for CNA analysis.

analyze_peaks(
  x,
  karyotypes = c("1:0", "1:1", "2:1", "2:0", "2:2"),
  min_karyotype_size = 0,
  min_absolute_karyotype_mutations = 10,
  p_binsize_peaks = 0.005,
  matching_epsilon = 0.03,
  n_bootstrap = 1,
  kernel_adjust = 1,
  matching_strategy = "closest"
)

Arguments

x
An object of class cnaqc, created by the init function.

karyotypes
The list of karyotypes to test. By default LOH regions (A, AA), diploid regions (AB), and amplification regions (AAB, AABB). These correspond to '1:0', '1:1', '2:1', '2:0', '2:2' in "Major:minor" notation.

min_karyotype_size
Karyotypes are subset by their size (normalized for the number of input mutations). Karyotypes smaller than this value are removed from analysis.

min_absolute_karyotype_mutations
As min_karyotype_size this imposes a minimum number of mutations on a karyotype to be analysed (in absolute number of mutations). Karyotypes smaller than this value are removed from analysis.

p_binsize_peaks
Peaks detected will be filtered if, in a peak, we map less than p_binsize_peaks * N mutations. The value N is obtained couting all mutations that map in all peaks.

matching_epsilon
Peaks at a location are matched with a range of epsilon * 0.5, left and right. By default (0.025) a maxium 5% tolerance is adopted. The status of overall PASS/ FAIL for this analysis is determined with the same tolerance, given the score computed across all karyotypes.

n_bootstrap
Number of times peak detection is bootstrapped (this helps sometimes finding peaks that might be visually observable but fail to be detected by the underlying peak detection heuristic). The default of this parameter is 1, meaning that no bootstrap is performed.

kernel_adjust
KDE adjust density parameter; see density. A Gaussian kernel is used (kernel = 'gaussian').

matching_strategy
If "closest", the closest peak will be used to match the expected peak. If "rightmost", peaks are matched prioritizing right to left peaks (the higher get matched first); this strategy is more correct in principle but works only if there are no spurious peaks in the estimated density. By defaule the first strategy is used.

Value

An object of class cnaqc, modified to hold the results from this analysis. See the vignette to see how to extract and plot the results.

Examples

data('example_dataset_CNAqc', package = 'CNAqc')
x = init(example_dataset_CNAqc$snvs, example_dataset_CNAqc$cna,example_dataset_CNAqc$purity)

x = analyze_peaks(x)
print(x)

print(x$peaks_analysis)

x = analyze_peaks(x, matching_strategy = "rightmost")
print(x)