Analyze CNA calls by peak detection.

analyze_peaks(x, karyotypes = c("1:0", "1:1", "2:1", "2:0", "2:2"),
  min_karyotype_size = 0.05, adjust = 1, min_density = 0.5,
  neighlim = 1, matching_epsilon = 0.015, ...)

Arguments

x
An object of class cnaqc, created by the init function.

karyotypes
The list of karyotypes to test. By default LOH regions (A, AA), diploid regions (AB), and amplification regions (AAB, AABB). These correspond to '1:0', '1:1', '2:1', '2:0', '2:2' in Major:minor notation.

min_karyotype_size
Karyotypes are subset by their size (normalized for the number of input mutations). Karyotypes smaller than this value are removed from analysis.

adjust
KDE adjust density parameter - see ?density. A Gaussian kernel is used (kernel = 'gaussian'). Default is 5

min_densityPeaks in the VAF distribution below this density value will not be used to match the expected peaks. This is used to remove low-density peaks due to noise and miscalled CNAs.

neighlimParameter for the peakPick package - see ?peakPick.

matching_epsilonPeaks at location \(x\) are matched in \(x-e\), \(x+e\) (inclusive), where \(e\) is this parameter. By default (0.015) a 3 is adopted.

...Parameters forwarded to a call to function peakpick from the peakPick package.