Analyze calls by peak detection.

This function carries out a peak-detection analysis to determine QC for clonal simple CNAs, clonal complex CNAs and subclonal CNAs.

Several scores are produced and returned; functions to visualize results are also available (see the tool website for example usages).

analyze_peaks(
  x,
  karyotypes = c("1:0", "1:1", "2:0", "2:1", "2:2"),
  min_karyotype_size = 0,
  min_absolute_karyotype_mutations = 100,
  p_binsize_peaks = 0.005,
  matching_epsilon = NULL,
  purity_error = 0.05,
  VAF_tolerance = 0.015,
  n_bootstrap = 1,
  kernel_adjust = 1,
  matching_strategy = "closest",
  KDE = TRUE,
  starting_state_subclonal_evolution = "1:1"
)

Arguments

x

An object of class cnaqc, created by the init function.

karyotypes

The list of karyotypes to test. By default LOH regions (A, AA), diploid regions (AB), and amplification regions (AAB, AABB). These correspond to '1:0', '1:1', '2:1', '2:0', '2:2' in "Major:minor" notation.

min_karyotype_size

Karyotypes are subset by their size (normalized for the number of input mutations). Karyotypes smaller than this value are removed from analysis.

min_absolute_karyotype_mutations

As min_karyotype_size this imposes a minimum number of mutations on a karyotype to be analysed (in absolute number of mutations). Karyotypes smaller than this value are removed from analysis.

p_binsize_peaks

Peaks detected will be filtered if, in a peak, we map less than p_binsize_peaks * N mutations. The value N is obtained couting all mutations that map in all peaks.#'

matching_epsilon

Deprecated parameter.

purity_error

Purity error tolerance to determine QC results. This can be set automatically using function auto_tolerance

VAF_tolerance

A tolerance in comparing bands overlaps - this is applied to the raw VAF values.

n_bootstrap

Number of times peak detection is bootstrapped (this helps sometimes finding peaks that might be visually observable but fail to be detected by the underlying peak detection heuristic). The default of this parameter is 1, meaning that no bootstrap is performed.

kernel_adjust

KDE adjust density parameter; see density. A Gaussian kernel is used (kernel = 'gaussian').

matching_strategy

If "closest", the closest peak will be used to match the expected peak. If "rightmost", peaks are matched prioritizing right to left peaks (the higher get matched first); this strategy is more correct in principle but works only if there are no spurious peaks in the estimated density. By defaule the first strategy is used.

KDE

Deprecated parameter (unused).

starting_state_subclonal_evolution

Starting state to determine linear versus branching evolutionary model, for sublonal CNA QC.

Value

An object of class cnaqc, modified to hold the results from this analysis. See the vignette to see how to extract and plot the results.

Examples

data('example_dataset_CNAqc', package = 'CNAqc')
x = init(example_dataset_CNAqc$snvs, example_dataset_CNAqc$cna,example_dataset_CNAqc$purity)
#>  [ CNAqc - CNA Quality Check ] 
#> 
#> ℹ Using reference genome coordinates for: GRCh38.
#> ℹ Drivers are annotated, but 'gene' column is missing, using mutation location.
#> ✔ Fortified calls for 12963 somatic mutations: 12963 SNVs (100%) and 0 indels.
#> ! CNAs have no CCF, assuming clonal CNAs (CCF = 1).
#> ✔ Fortified CNAs for 267 segments: 267 clonal and 0 subclonal.
#> ✔ 12963 mutations mapped to clonal CNAs.

x = analyze_peaks(x)
#> 
#> ── Peak analysis: simple CNAs ──────────────────────────────────────────────────
#> 
#> ℹ Analysing 9041 mutations mapping to karyotype(s) 2:2 and 2:1.
#> ℹ Mixed type peak detection for karyotype 2:1 (1563 mutations)
#> ℹ Mixed type peak detection for karyotype 2:2 (7478 mutations)
#> # A tibble: 4 × 16
#> # Rowwise: 
#>   mutation_mu…¹ karyo…²  peak delta…³     x     y count…⁴ disca…⁵ from  offset…⁶
#>           <dbl> <chr>   <dbl>   <dbl> <dbl> <dbl>   <int> <lgl>   <chr>    <dbl>
#> 1             1 2:2     0.235 0.00700 0.24   0.86      65 FALSE   KDE   -4.55e-3
#> 2             2 2:2     0.471 0.0140  0.46   8.14     625 FALSE   KDE    1.09e-2
#> 3             1 2:1     0.308 0.0120  0.307  3.60      59 FALSE   BMix   6.56e-4
#> 4             2 2:1     0.616 0.0239  0.602  2.94      60 FALSE   BMix   1.38e-2
#> # … with 6 more variables: offset <dbl>, weight <dbl>, epsilon <dbl>,
#> #   VAF_tolerance <dbl>, matched <lgl>, QC <chr>, and abbreviated variable
#> #   names ¹​mutation_multiplicity, ²​karyotype, ³​delta_vaf, ⁴​counts_per_bin,
#> #   ⁵​discarded, ⁶​offset_VAF
#> ✔ Peak detection PASS with r = 0.00844850061748855 - maximum purity error ε = 0.05.
#> 
#> ── Peak analysis: complex CNAs ─────────────────────────────────────────────────
#> 
#> ℹ Karyotypes 4:2, 3:2, and 3:0 with >100 mutation(s). Using epsilon = 0.05.
#> # A tibble: 3 × 5
#> # Groups:   karyotype, matched [3]
#>   karyotype n       matched mismatched  prop
#>   <chr>     <table>   <int>      <dbl> <dbl>
#> 1 3:0        312          3          0     1
#> 2 3:2       1625          3          0     1
#> 3 4:2       1893          4          0     1
#> 
#> ── Peak analysis: subclonal CNAs ───────────────────────────────────────────────
#> 
#> ℹ No subclonal CNAs in this sample.
print(x)
#> ── [ CNAqc ]  12963 mutations in 267 segments (267 clonal, 0 subclonal). Genome 
#> 
#> ── Clonal CNAs 
#> 
#>    2:2  [n = 7478, L = 1483 Mb] ■■■■■■■■■■■■■■■■■■■■■■■■■■■  { CTCF }
#>    4:2  [n = 1893, L =  331 Mb] ■■■■■■■
#>    3:2  [n = 1625, L =  357 Mb] ■■■■■■
#>    2:1  [n = 1563, L =  420 Mb] ■■■■■■  { TTN }
#>    3:0  [n =  312, L =  137 Mb] ■
#>    2:0  [n =   81, L =   39 Mb]   { TP53 }
#>   16:2  [n =    4, L =    0 Mb] 
#>   25:2  [n =    2, L =    1 Mb] 
#>    3:1  [n =    2, L =    1 Mb] 
#>  106:1  [n =    1, L =    0 Mb] 
#> 
#> ℹ Sample Purity: 89% ~ Ploidy: 4.
#> 
#> ℹ There are 3 annotated driver(s) mapped to clonal CNAs.
#>          chr      from        to ref alt  DP NV       VAF driver_label is_driver
#>         chr2 179431633 179431634   C   T 117 77 0.6581197          TTN      TRUE
#>        chr16  67646006  67646007   C   T 120 54 0.4500000         CTCF      TRUE
#>        chr17   7577106   7577107   G   C  84 78 0.9285714         TP53      TRUE
#> 
#> ──  PASS  Peaks QC closest: 200%, λ = 0.0084. Purity correction: 1%. ───────────
#> 
#> ℹ 2:2 ~ n = 7478  ( 83%) →  PASS  -0.034   PASS  0.037  
#> ℹ 2:1 ~ n = 1563  ( 17%) →  PASS  0.003    PASS  0.028  
#> 
#> ── General peak QC (3830 mutations):  PASS  10  FAIL  0 - epsilon = 0.05. ──────
#> 
#> ℹ 3:0 ~ n = 312   (  8%) →  PASS  3  FAIL  0
#> ℹ 3:2 ~ n = 1625  ( 42%) →  PASS  3  FAIL  0
#> ℹ 4:2 ~ n = 1893  ( 49%) →  PASS  4  FAIL  0

print(x$peaks_analysis)
#> $score
#> [1] 0.008448501
#> 
#> $fits
#> $fits$`2:1`
#> $fits$`2:1`$xy_peaks
#> # A tibble: 6 × 5
#>       x     y counts_per_bin discarded from 
#>   <dbl> <dbl>          <int> <lgl>     <chr>
#> 1 0.08  0.27               5 FALSE     KDE  
#> 2 0.3   3.6               78 FALSE     KDE  
#> 3 0.6   2.94              60 FALSE     KDE  
#> 4 0.307 3.60              59 FALSE     BMix 
#> 5 0.602 2.94              60 FALSE     BMix 
#> 6 0.111 0.243              2 FALSE     BMix 
#> 
#> $fits$`2:1`$density
#> 
#> Call:
#> 	density.default(x = y, adjust = kernel_adjust, kernel = "gaussian",     na.rm = T)
#> 
#> Data: y (1563 obs.);	Bandwidth 'bw' = 0.0341
#> 
#>        x                  y          
#>  Min.   :-0.05578   Min.   :0.00016  
#>  1st Qu.: 0.17384   1st Qu.:0.21293  
#>  Median : 0.40346   Median :0.51690  
#>  Mean   : 0.40346   Mean   :1.08768  
#>  3rd Qu.: 0.63308   3rd Qu.:1.99965  
#>  Max.   : 0.86270   Max.   :3.60365  
#> 
#> $fits$`2:1`$matching
#> # A tibble: 2 × 16
#> # Rowwise: 
#>   mutation_mul…¹ karyo…²  peak delta…³     x     y count…⁴ disca…⁵ from  offse…⁶
#>            <dbl> <chr>   <dbl>   <dbl> <dbl> <dbl>   <int> <lgl>   <chr>   <dbl>
#> 1              1 2:1     0.308  0.0120 0.307  3.60      59 FALSE   BMix  6.56e-4
#> 2              2 2:1     0.616  0.0239 0.602  2.94      60 FALSE   BMix  1.38e-2
#> # … with 6 more variables: offset <dbl>, weight <dbl>, epsilon <dbl>,
#> #   VAF_tolerance <dbl>, matched <lgl>, QC <chr>, and abbreviated variable
#> #   names ¹​mutation_multiplicity, ²​karyotype, ³​delta_vaf, ⁴​counts_per_bin,
#> #   ⁵​discarded, ⁶​offset_VAF
#> 
#> 
#> $fits$`2:2`
#> $fits$`2:2`$xy_peaks
#> # A tibble: 19 × 5
#>         x     y counts_per_bin discarded from 
#>     <dbl> <dbl>          <int> <lgl>     <chr>
#>  1 0.06   0.94              89 FALSE     KDE  
#>  2 0.11   0.27              20 TRUE      KDE  
#>  3 0.22   0.84              69 FALSE     KDE  
#>  4 0.24   0.86              65 FALSE     KDE  
#>  5 0.46   8.14             625 FALSE     KDE  
#>  6 0.67   0.01               0 TRUE      KDE  
#>  7 0.72   0.02               2 TRUE      KDE  
#>  8 0.75   0.01               2 TRUE      KDE  
#>  9 0.79   0.01               1 TRUE      KDE  
#> 10 0.82   0.01               1 TRUE      KDE  
#> 11 0.85   0.01               1 TRUE      KDE  
#> 12 0.91   0.01               1 TRUE      KDE  
#> 13 0.93   0.01               0 TRUE      KDE  
#> 14 0.97   0.02               0 TRUE      KDE  
#> 15 1      0.02               3 TRUE      KDE  
#> 16 0.221  0.832             69 FALSE     BMix 
#> 17 0.482  7.02             606 FALSE     BMix 
#> 18 0.0740 0.558             57 FALSE     BMix 
#> 19 0.439  6.65             477 FALSE     BMix 
#> 
#> $fits$`2:2`$density
#> 
#> Call:
#> 	density.default(x = y, adjust = kernel_adjust, kernel = "gaussian",     na.rm = T)
#> 
#> Data: y (7478 obs.);	Bandwidth 'bw' = 0.008033
#> 
#>        x                 y           
#>  Min.   :0.02101   Min.   :0.000244  
#>  1st Qu.:0.27179   1st Qu.:0.007045  
#>  Median :0.52256   Median :0.203133  
#>  Mean   :0.52256   Mean   :0.995946  
#>  3rd Qu.:0.77333   3rd Qu.:0.787272  
#>  Max.   :1.02410   Max.   :8.138867  
#> 
#> $fits$`2:2`$matching
#> # A tibble: 2 × 16
#> # Rowwise: 
#>   mutation_mu…¹ karyo…²  peak delta…³     x     y count…⁴ disca…⁵ from  offset…⁶
#>           <dbl> <chr>   <dbl>   <dbl> <dbl> <dbl>   <int> <lgl>   <chr>    <dbl>
#> 1             1 2:2     0.235 0.00700  0.24  0.86      65 FALSE   KDE   -0.00455
#> 2             2 2:2     0.471 0.0140   0.46  8.14     625 FALSE   KDE    0.0109 
#> # … with 6 more variables: offset <dbl>, weight <dbl>, epsilon <dbl>,
#> #   VAF_tolerance <dbl>, matched <lgl>, QC <chr>, and abbreviated variable
#> #   names ¹​mutation_multiplicity, ²​karyotype, ³​delta_vaf, ⁴​counts_per_bin,
#> #   ⁵​discarded, ⁶​offset_VAF
#> 
#> 
#> 
#> $matches
#> # A tibble: 4 × 16
#> # Rowwise: 
#>   mutation_mu…¹ karyo…²  peak delta…³     x     y count…⁴ disca…⁵ from  offset…⁶
#>           <dbl> <chr>   <dbl>   <dbl> <dbl> <dbl>   <int> <lgl>   <chr>    <dbl>
#> 1             1 2:2     0.235 0.00700 0.24   0.86      65 FALSE   KDE   -4.55e-3
#> 2             2 2:2     0.471 0.0140  0.46   8.14     625 FALSE   KDE    1.09e-2
#> 3             1 2:1     0.308 0.0120  0.307  3.60      59 FALSE   BMix   6.56e-4
#> 4             2 2:1     0.616 0.0239  0.602  2.94      60 FALSE   BMix   1.38e-2
#> # … with 6 more variables: offset <dbl>, weight <dbl>, epsilon <dbl>,
#> #   VAF_tolerance <dbl>, matched <lgl>, QC <chr>, and abbreviated variable
#> #   names ¹​mutation_multiplicity, ²​karyotype, ³​delta_vaf, ⁴​counts_per_bin,
#> #   ⁵​discarded, ⁶​offset_VAF
#> 
#> $matching_strategy
#> [1] "closest"
#> 
#> $purity_error
#> [1] 0.05
#> 
#> $min_karyotype_size
#> [1] 0
#> 
#> $p_binsize_peaks
#> [1] 0.005
#> 
#> $QC
#> [1] "PASS"
#> 
#> $KDE
#> [1] NA
#> 
#> $general
#> $general$analysis
#> [1] "3:2" "3:0" "4:2"
#> 
#> $general$params
#> $general$params$n_min
#> [1] 100
#> 
#> $general$params$epsilon
#> [1] 0.05
#> 
#> $general$params$n_bootstrap
#> [1] 1
#> 
#> 
#> $general$expected_peaks
#> # A tibble: 10 × 9
#>    minor Major ploidy multiplicity purity  peak karyotype matched n      
#>    <dbl> <dbl>  <dbl>        <int>  <dbl> <dbl> <chr>     <lgl>   <table>
#>  1     2     3      5            1   0.89 0.191 3:2       TRUE    1625   
#>  2     2     3      5            2   0.89 0.381 3:2       TRUE    1625   
#>  3     2     3      5            3   0.89 0.572 3:2       TRUE    1625   
#>  4     0     3      3            1   0.89 0.308 3:0       TRUE     312   
#>  5     0     3      3            2   0.89 0.616 3:0       TRUE     312   
#>  6     0     3      3            3   0.89 0.924 3:0       TRUE     312   
#>  7     2     4      6            1   0.89 0.160 4:2       TRUE    1893   
#>  8     2     4      6            2   0.89 0.320 4:2       TRUE    1893   
#>  9     2     4      6            3   0.89 0.480 4:2       TRUE    1893   
#> 10     2     4      6            4   0.89 0.640 4:2       TRUE    1893   
#> 
#> $general$data_peaks
#> # A tibble: 20 × 7
#>        x     y counts_per_bin discarded from  karyotype n      
#>    <dbl> <dbl>          <int> <lgl>     <chr> <chr>     <table>
#>  1  0.07  1.33              6 FALSE     KDE   3:0        312   
#>  2  0.17  0.39              1 TRUE      KDE   3:0        312   
#>  3  0.31  0.72              2 FALSE     KDE   3:0        312   
#>  4  0.42  0.08              0 TRUE      KDE   3:0        312   
#>  5  0.48  0.08              0 TRUE      KDE   3:0        312   
#>  6  0.63  0.42              2 TRUE      KDE   3:0        312   
#>  7  0.76  0.2               2 TRUE      KDE   3:0        312   
#>  8  0.92 10.6              34 FALSE     KDE   3:0        312   
#>  9  0.18  2.13             40 FALSE     KDE   3:2       1625   
#> 10  0.37  4.39             94 FALSE     KDE   3:2       1625   
#> 11  0.56  1.61             33 FALSE     KDE   3:2       1625   
#> 12  0.06  0.89             18 FALSE     KDE   4:2       1893   
#> 13  0.15  1.62             30 FALSE     KDE   4:2       1893   
#> 14  0.31  7.93            158 FALSE     KDE   4:2       1893   
#> 15  0.49  0.03              1 TRUE      KDE   4:2       1893   
#> 16  0.64  1.65             34 FALSE     KDE   4:2       1893   
#> 17  0.85  0                 0 TRUE      KDE   4:2       1893   
#> 18  0.86  0                 0 TRUE      KDE   4:2       1893   
#> 19  0.88  0                 0 TRUE      KDE   4:2       1893   
#> 20  1     0.02              1 TRUE      KDE   4:2       1893   
#> 
#> $general$data_densities
#> # A tibble: 1,536 × 4
#>          x       y karyotype n      
#>      <dbl>   <dbl> <chr>     <table>
#>  1 0.00120 0.00319 3:0       312    
#>  2 0.00325 0.00479 3:0       312    
#>  3 0.00530 0.00706 3:0       312    
#>  4 0.00735 0.0102  3:0       312    
#>  5 0.00940 0.0146  3:0       312    
#>  6 0.0115  0.0204  3:0       312    
#>  7 0.0135  0.0282  3:0       312    
#>  8 0.0155  0.0383  3:0       312    
#>  9 0.0176  0.0512  3:0       312    
#> 10 0.0196  0.0675  3:0       312    
#> # … with 1,526 more rows
#> 
#> $general$summary
#> # A tibble: 3 × 5
#> # Groups:   karyotype, matched [3]
#>   karyotype n       matched mismatched  prop
#>   <chr>     <table>   <int>      <dbl> <dbl>
#> 1 3:0        312          3          0     1
#> 2 3:2       1625          3          0     1
#> 3 4:2       1893          4          0     1
#> 
#> 

x = analyze_peaks(x, matching_strategy = "rightmost")
#> 
#> ── Peak analysis: simple CNAs ──────────────────────────────────────────────────
#> 
#> ℹ Analysing 9041 mutations mapping to karyotype(s) 2:2 and 2:1.
#> ℹ Mixed type peak detection for karyotype 2:1 (1563 mutations)
#> ℹ Mixed type peak detection for karyotype 2:2 (7478 mutations)
#> # A tibble: 4 × 16
#> # Rowwise: 
#>   mutation_mu…¹ karyo…²  peak delta…³     x     y count…⁴ disca…⁵ from  offset…⁶
#>           <dbl> <chr>   <dbl>   <dbl> <dbl> <dbl>   <int> <lgl>   <chr>    <dbl>
#> 1             1 2:2     0.235 0.00700 0.24   0.86      65 FALSE   KDE   -0.00455
#> 2             2 2:2     0.471 0.0140  0.461  8.14     625 FALSE   BMix   0.0102 
#> 3             1 2:1     0.308 0.0120  0.306  3.60      59 FALSE   BMix   0.00245
#> 4             2 2:1     0.616 0.0239  0.602  2.94      60 FALSE   BMix   0.0138 
#> # … with 6 more variables: offset <dbl>, weight <dbl>, epsilon <dbl>,
#> #   VAF_tolerance <dbl>, matched <lgl>, QC <chr>, and abbreviated variable
#> #   names ¹​mutation_multiplicity, ²​karyotype, ³​delta_vaf, ⁴​counts_per_bin,
#> #   ⁵​discarded, ⁶​offset_VAF
#> ✔ Peak detection PASS with r = 0.00776282580370826 - maximum purity error ε = 0.05.
#> 
#> ── Peak analysis: complex CNAs ─────────────────────────────────────────────────
#> 
#> ℹ Karyotypes 4:2, 3:2, and 3:0 with >100 mutation(s). Using epsilon = 0.05.
#> # A tibble: 3 × 5
#> # Groups:   karyotype, matched [3]
#>   karyotype n       matched mismatched  prop
#>   <chr>     <table>   <int>      <dbl> <dbl>
#> 1 3:0        312          3          0     1
#> 2 3:2       1625          3          0     1
#> 3 4:2       1893          4          0     1
#> 
#> ── Peak analysis: subclonal CNAs ───────────────────────────────────────────────
#> 
#> ℹ No subclonal CNAs in this sample.
print(x)
#> ── [ CNAqc ]  12963 mutations in 267 segments (267 clonal, 0 subclonal). Genome 
#> 
#> ── Clonal CNAs 
#> 
#>    2:2  [n = 7478, L = 1483 Mb] ■■■■■■■■■■■■■■■■■■■■■■■■■■■  { CTCF }
#>    4:2  [n = 1893, L =  331 Mb] ■■■■■■■
#>    3:2  [n = 1625, L =  357 Mb] ■■■■■■
#>    2:1  [n = 1563, L =  420 Mb] ■■■■■■  { TTN }
#>    3:0  [n =  312, L =  137 Mb] ■
#>    2:0  [n =   81, L =   39 Mb]   { TP53 }
#>   16:2  [n =    4, L =    0 Mb] 
#>   25:2  [n =    2, L =    1 Mb] 
#>    3:1  [n =    2, L =    1 Mb] 
#>  106:1  [n =    1, L =    0 Mb] 
#> 
#> ℹ Sample Purity: 89% ~ Ploidy: 4.
#> 
#> ℹ There are 3 annotated driver(s) mapped to clonal CNAs.
#>          chr      from        to ref alt  DP NV       VAF driver_label is_driver
#>         chr2 179431633 179431634   C   T 117 77 0.6581197          TTN      TRUE
#>        chr16  67646006  67646007   C   T 120 54 0.4500000         CTCF      TRUE
#>        chr17   7577106   7577107   G   C  84 78 0.9285714         TP53      TRUE
#> 
#> ──  PASS  Peaks QC closest: 200%, λ = 0.0078. Purity correction: 1%. ───────────
#> 
#> ℹ 2:2 ~ n = 7478  ( 83%) →  PASS  -0.034   PASS  0.035  
#> ℹ 2:1 ~ n = 1563  ( 17%) →  PASS  0.01     PASS  0.028  
#> 
#> ── General peak QC (3830 mutations):  PASS  10  FAIL  0 - epsilon = 0.05. ──────
#> 
#> ℹ 3:0 ~ n = 312   (  8%) →  PASS  3  FAIL  0
#> ℹ 3:2 ~ n = 1625  ( 42%) →  PASS  3  FAIL  0
#> ℹ 4:2 ~ n = 1893  ( 49%) →  PASS  4  FAIL  0