Analyze CNA calls by peak detection.

This function carries out a peak-detection analysis based on KDE and the package peakPick in order to determine if the mutations that map to a certain karyotype have an allelic frequency that is consistent with the Major and minor alleles of the segment, and the tumour purity. A score is produced as a linear combination of the distance of the actual peak to the expected one, derived with standard equations for CNA analysis.

analyze_peaks(x, karyotypes = c("1:0", "1:1", "2:1", "2:0", "2:2"),
  min_karyotype_size = 0.05, p_binsize_peaks = 0.005,
  matching_epsilon = 0.025, kernel_adjust = 1)

Arguments

x
An object of class cnaqc, created by the init function.

karyotypes
The list of karyotypes to test. By default LOH regions (A, AA), diploid regions (AB), and amplification regions (AAB, AABB). These correspond to '1:0', '1:1', '2:1', '2:0', '2:2' in "Major:minor" notation.

min_karyotype_size
Karyotypes are subset by their size (normalized for the number of input mutations). Karyotypes smaller than this value are removed from analysis.

p_binsize_peaks
Peaks detected will be filtered if, in a peak, we map less than p_binsize_peaks * N mutations. The value N is obtained couting all mutations that map in all peaks.

matching_epsilon
Peaks at a location are matched with a range of epsilon * 0.5, left and right. By default (0.025) a maxium 5% tolerance is adopted. The status of overall PASS/ FAIL for this analysis is determined with the same tolerance, given the score computed across all karyotypes.

kernel_adjust
KDE adjust density parameter; see density. A Gaussian kernel is used (kernel = 'gaussian').

Value

An object of class cnaqc, modified to hold the results from this analysis. See the vignette to see how to extract and plot the results.

Examples

data('example_dataset_CNAqc', package = 'CNAqc')
x = init(example_dataset_CNAqc$snvs, example_dataset_CNAqc$cna,example_dataset_CNAqc$purity)

#>  [ CNAqc - CNA Quality Check ] 
#> 

#> ℹ Using reference genome coordinates for: GRCh38.

#> ℹ Drivers are annotated, but 'gene' column is missing, using mutation location.

#> ! Missing CCF column from CNA calls, adding CCF = 1 assuming clonal CNA calls.

#> ℹ Input n = 13141 mutations for 267 CNA segments (267 clonal, 0 subclonal)

#> 

#> ✓ Mapped n = 12963 mutations to clonal segments (99% of input)

x = analyze_peaks(x)

#> ℹ Requested karyotypes 1:0, 1:1, 2:1, 2:0 and 2:2.

#> ℹ Found n = 9041 mutations in 2:2, 2:1 (skipping those with n < 657 mutations).

#> # A tibble: 4 x 12
#>   mutation_multip… karyotype  peak     x     y counts_per_bin discarded   offset
#>              <dbl> <chr>     <dbl> <dbl> <dbl>          <int> <lgl>        <dbl>
#> 1                2 2:1       0.616  0.61  2.86             45 FALSE      0.00592
#> 2                1 2:1       0.308  0.31  3.73             72 FALSE     -0.00204
#> 3                2 2:2       0.471  0.46  8.16            612 FALSE      0.0109 
#> 4                1 2:2       0.235  0.25  0.75             70 FALSE     -0.0146 
#> # … with 4 more variables: matched <lgl>, weight <dbl>, score <dbl>, QC <chr>

#> ✓ Peak detection PASS with r = -0.00117483319779741 and tolerance e = 0.05

print(x)

#> ── [ CNAqc ]  n = 13141 mutations in 267 segments (267 clonal + 0 subclonal). Genome reference: GRCh38. ──────────────────

#> 
#>    2:2  [n = 7478] ■■■■■■■■■■■■■■■■■■■■■■■■■■■  { chr19:15292495:15292496:G:A }
#>    4:2  [n = 1893] ■■■■■■■
#>    3:2  [n = 1625] ■■■■■■
#>    2:1  [n = 1563] ■■■■■■  { chr2:179431633:179431634:C:T }
#>    3:0  [n =  312] ■
#> 

#> ℹ Sample Purity: 89% ~ Ploidy: 4.

#> 

#> ℹ Mutations annotated have 3 drivers.

#> ✓ QC via peak detection available, score: -0.00117483319779741.

#> # A tibble: 4 x 12
#>   mutation_multip… karyotype  peak     x     y counts_per_bin discarded   offset
#>              <dbl> <chr>     <dbl> <dbl> <dbl>          <int> <lgl>        <dbl>
#> 1                2 2:1       0.616  0.61  2.86             45 FALSE      0.00592
#> 2                1 2:1       0.308  0.31  3.73             72 FALSE     -0.00204
#> 3                2 2:2       0.471  0.46  8.16            612 FALSE      0.0109 
#> 4                1 2:2       0.235  0.25  0.75             70 FALSE     -0.0146 
#> # … with 4 more variables: matched <lgl>, weight <dbl>, score <dbl>, QC <chr>

print(x$peaks_analysis)

#> $score
#> [1] -0.001174833
#> 
#> $fits
#> $fits$`1:0`
#> $fits$`1:0`$matching
#> NULL
#> 
#> 
#> $fits$`1:1`
#> $fits$`1:1`$matching
#> NULL
#> 
#> 
#> $fits$`2:1`
#> $fits$`2:1`$matching
#> # A tibble: 2 x 10
#>   mutation_multip… karyotype  peak     x     y counts_per_bin discarded   offset
#>              <dbl> <chr>     <dbl> <dbl> <dbl>          <int> <lgl>        <dbl>
#> 1                2 2:1       0.616  0.61  2.86             45 FALSE      0.00592
#> 2                1 2:1       0.308  0.31  3.73             72 FALSE     -0.00204
#> # … with 2 more variables: matched <lgl>, weight <dbl>
#> 
#> $fits$`2:1`$density
#> 
#> Call:
#> 	density.default(x = y, adjust = kernel_adjust, kernel = "gaussian")
#> 
#> Data: y (1563 obs.);	Bandwidth 'bw' = 0.03344
#> 
#>        x                  y          
#>  Min.   :-0.05113   Min.   :0.00017  
#>  1st Qu.: 0.17683   1st Qu.:0.13437  
#>  Median : 0.40480   Median :0.54953  
#>  Mean   : 0.40480   Mean   :1.09559  
#>  3rd Qu.: 0.63276   3rd Qu.:1.98170  
#>  Max.   : 0.86073   Max.   :3.73431  
#> 
#> $fits$`2:1`$xy_peaks
#> # A tibble: 2 x 4
#>       x     y counts_per_bin discarded
#>   <dbl> <dbl>          <int> <lgl>    
#> 1  0.31  3.73             72 FALSE    
#> 2  0.61  2.86             45 FALSE    
#> 
#> 
#> $fits$`2:0`
#> $fits$`2:0`$matching
#> NULL
#> 
#> 
#> $fits$`2:2`
#> $fits$`2:2`$matching
#> # A tibble: 2 x 10
#>   mutation_multip… karyotype  peak     x     y counts_per_bin discarded  offset
#>              <dbl> <chr>     <dbl> <dbl> <dbl>          <int> <lgl>       <dbl>
#> 1                2 2:2       0.471  0.46  8.16            612 FALSE      0.0109
#> 2                1 2:2       0.235  0.25  0.75             70 FALSE     -0.0146
#> # … with 2 more variables: matched <lgl>, weight <dbl>
#> 
#> $fits$`2:2`$density
#> 
#> Call:
#> 	density.default(x = y, adjust = kernel_adjust, kernel = "gaussian")
#> 
#> Data: y (7478 obs.);	Bandwidth 'bw' = 0.00801
#> 
#>        x                 y           
#>  Min.   :0.02108   Min.   :0.000000  
#>  1st Qu.:0.27182   1st Qu.:0.008399  
#>  Median :0.52256   Median :0.181167  
#>  Mean   :0.52256   Mean   :0.996088  
#>  3rd Qu.:0.77329   3rd Qu.:0.717005  
#>  Max.   :1.02403   Max.   :8.156854  
#> 
#> $fits$`2:2`$xy_peaks
#> # A tibble: 9 x 4
#>       x     y counts_per_bin discarded
#>   <dbl> <dbl>          <int> <lgl>    
#> 1  0.11  0.24             16 TRUE     
#> 2  0.21  0.85             57 FALSE    
#> 3  0.25  0.75             70 FALSE    
#> 4  0.46  8.16            612 FALSE    
#> 5  0.67  0.01              0 TRUE     
#> 6  0.73  0.02              0 TRUE     
#> 7  0.85  0.01              1 TRUE     
#> 8  0.91  0.02              2 TRUE     
#> 9  0.97  0.03              0 TRUE     
#> 
#> 
#> 
#> $matches
#> # A tibble: 4 x 12
#>   mutation_multip… karyotype  peak     x     y counts_per_bin discarded   offset
#>              <dbl> <chr>     <dbl> <dbl> <dbl>          <int> <lgl>        <dbl>
#> 1                2 2:1       0.616  0.61  2.86             45 FALSE      0.00592
#> 2                1 2:1       0.308  0.31  3.73             72 FALSE     -0.00204
#> 3                2 2:2       0.471  0.46  8.16            612 FALSE      0.0109 
#> 4                1 2:2       0.235  0.25  0.75             70 FALSE     -0.0146 
#> # … with 4 more variables: matched <lgl>, weight <dbl>, score <dbl>, QC <chr>
#> 
#> $min_karyotype_size
#> [1] 0.05
#> 
#> $p_binsize_peaks
#> [1] 0.005
#> 
#> $matching_epsilon
#> [1] 0.025
#> 
#> $QC
#> [1] "PASS"
#>