Annotate variant in a CNAqc object

The function provides an easy and fast way to annotate SNVs, in particular we annotate the locations of the different mutations, the consequences for those in coding regions and we compare them to a set of known drivers.

annotate_variants(
  x,
  ref = "hg19",
  driver_list = CNAqc::intogen_drivers,
  polyphen = FALSE,
  filter_tumor_type = NULL
)

Arguments

x
a data.frame with coloumns c("chr", "from", "to").

driver_list
a one coloumn data.frame with gene symbols

polyphen
whete to annotate polyphen scores

filter_tumor_type
filter intogen drivers by tumor type

Value

a data.frame with additiona coloumns `location` with the position of the variant in the genome (coding, intron, threeUTR, ...), `consequence` with the consequence of coding mutations (synonymous, nonsynonymous, ...), `is_driver` a bool that indicate if the gene is a driver, `gene_symbol` for the annotated corresponding gene symbol (if the variant is in a gene) and `driver_label` with the driver label written as gene_refAA->varAA (is NA in case `is_driver` is FALSE).

The `chr` coloumn int the input should contain the chr prefix before the chromosome number. The annotation process is based on the package VariantAnnotation check it for more information.

Examples

if (FALSE) library(CNAqc)

data('example_dataset_CNAqc', package = 'CNAqc')

snvs <- example_dataset_CNAqc$snvs

snvs_annotated <- annotate_variants(snvs)

#> 
#> Attaching package: ‘BiocGenerics’

#> The following objects are masked from ‘package:parallel’:
#> 
#>     clusterApply, clusterApplyLB, clusterCall, clusterEvalQ,
#>     clusterExport, clusterMap, parApply, parCapply, parLapply,
#>     parLapplyLB, parRapply, parSapply, parSapplyLB

#> The following objects are masked from ‘package:dplyr’:
#> 
#>     combine, intersect, setdiff, union

#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs

#> The following objects are masked from ‘package:base’:
#> 
#>     anyDuplicated, append, as.data.frame, basename, cbind, colnames,
#>     dirname, do.call, duplicated, eval, evalq, Filter, Find, get, grep,
#>     grepl, intersect, is.unsorted, lapply, Map, mapply, match, mget,
#>     order, paste, pmax, pmax.int, pmin, pmin.int, Position, rank,
#>     rbind, Reduce, rownames, sapply, setdiff, sort, table, tapply,
#>     union, unique, unsplit, which.max, which.min

#> 
#> Attaching package: ‘S4Vectors’

#> The following objects are masked from ‘package:dplyr’:
#> 
#>     first, rename

#> The following object is masked from ‘package:tidyr’:
#> 
#>     expand

#> The following object is masked from ‘package:base’:
#> 
#>     expand.grid

#> 
#> Attaching package: ‘IRanges’

#> The following objects are masked from ‘package:dplyr’:
#> 
#>     collapse, desc, slice

#> The following object is masked from ‘package:purrr’:
#> 
#>     reduce

#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.

#> 
#> Attaching package: ‘AnnotationDbi’

#> The following object is masked from ‘package:dplyr’:
#> 
#>     select

#> 
#> Attaching package: ‘XVector’

#> The following object is masked from ‘package:purrr’:
#> 
#>     compact

#> 
#> Attaching package: ‘Biostrings’

#> The following object is masked from ‘package:base’:
#> 
#>     strsplit

#> 

#> → Preparing annotations for a dataframe

#> ! There are existing annotations in your data, they will be cancelled.

#> # A tibble: 3 x 13
#>   chr        from       to ref   alt   FILTER    DP    NV   VAF ANNOVAR_FUNCTION
#>   <chr>     <dbl>    <dbl> <chr> <chr> <chr>  <dbl> <dbl> <dbl> <chr>           
#> 1 chr2  179431633   1.79e8 C     T     PASS     117    77 0.658 exonic          
#> 2 chr16  67646006   6.76e7 C     T     PASS     120    54 0.45  exonic          
#> 3 chr17   7577106   7.58e6 G     C     PASS      84    78 0.929 exonic          
#> # … with 3 more variables: GENE <chr>, is_driver <lgl>, driver_label <chr>

#> ℹ Locaring variants with VariantAnnotation

#> Warning: GRanges object contains 364 out-of-bound ranges located on sequences
#>   4529, 4631, 4632, 1653, 1654, 1655, 1656, 12192, 12362, 12363, 12364,
#>   12365, 12366, 12367, 12517, 12518, 12519, 12520, 13123, 13124, 13974,
#>   13975, 13976, 13994, 13995, 17007, 17008, 17009, 17010, 17011, 17012,
#>   17014, 17015, 17016, 17793, 18007, 20532, 22527, 22954, 22955, 22956,
#>   23347, 27275, 27276, 28075, 28085, 28893, 29046, 29047, 29242, 31362,
#>   32048, 32050, 33339, 33714, 33715, 32912, 32913, 40960, 40961, 44219,
#>   44220, 44242, 44262, 44263, 44356, 46050, 46054, 46062, 46063, 46064,
#>   46065, 46066, 53476, 53477, 53497, 53498, 59180, 57764, 57766, 61092,
#>   69815, 69930, 70330, 70331, 70332, 77129, 77130, 76239, 76240, 76241,
#>   77987, and 77988. Note that ranges located on a sequence whose length
#>   is unknown (NA) or on a circular sequence are not considered
#>   out-of-bound (use seqlengths() and isCircular() to get the lengths and
#>   circularity flags of the underlying sequences). You can use trim() to
#>   trim these ranges. See ?`trim,GenomicRanges-method` for more
#>   information.

#> 'select()' returned many:1 mapping between keys and columns

#> 'select()' returned many:1 mapping between keys and columns

#> 'select()' returned 1:1 mapping between keys and columns

#> 'select()' returned many:1 mapping between keys and columns

#> 'select()' returned 1:1 mapping between keys and columns

#> 'select()' returned many:1 mapping between keys and columns

#> ✔ Locaring variants with VariantAnnotation ... done

#> 

#> ℹ Sourcing organism

#> using temporary cache /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpPdUtgb/BiocFileCache

#> creating 'src_organism' database...

#> Warning: Ignoring remaining part of query: CREATE INDEX IF NOT EXISTS entrez_accession ON id_accession (entrez);

#> Warning: Ignoring remaining part of query: CREATE INDEX IF NOT EXISTS entrez_protein ON id_protein (entrez);

#> ✔ Sourcing organism ... done

#> 

#> ℹ Traslating Entrez ids

#> ✔ Traslating Entrez ids ... done

#> 

#> ℹ Transforming data

#> ✔ Transforming data ... done

#> 

#> ℹ Predicting coding

#> Warning: GRanges object contains 315 out-of-bound ranges located on sequences
#>   4529, 4631, 4632, 1653, 1654, 1655, 1656, 12192, 12362, 12363, 12364,
#>   12365, 12366, 12367, 12517, 12518, 12519, 12520, 13123, 13124, 13974,
#>   13975, 13976, 13994, 13995, 17007, 17008, 17009, 17010, 17011, 17012,
#>   17014, 17015, 17016, 17793, 18007, 20532, 22527, 22954, 22955, 22956,
#>   23347, 27275, 27276, 28075, 28085, 28893, 29046, 29047, 29242, 31362,
#>   32048, 32050, 33339, 33714, 33715, 32912, 32913, 40960, 40961, 44219,
#>   44220, 44242, 44262, 44263, 44356, 53476, 53477, 53497, 53498, 59180,
#>   57764, 57766, 61092, 69815, 69930, 70330, 70331, 70332, 77129, 77130,
#>   76239, 76240, 76241, 77987, and 77988. Note that ranges located on a
#>   sequence whose length is unknown (NA) or on a circular sequence are not
#>   considered out-of-bound (use seqlengths() and isCircular() to get the
#>   lengths and circularity flags of the underlying sequences). You can use
#>   trim() to trim these ranges. See ?`trim,GenomicRanges-method` for more
#>   information.

#> ✔ Predicting coding ... done

#> 

#> 

#> ── Found 3 driver(s) 

#> # A tibble: 3 x 10
#>   chr      from     to gene_symbol location consequence refAA varAA driver_label
#>   <fct>   <int>  <int> <chr>       <chr>    <chr>       <chr> <chr> <chr>       
#> 1 chr16  6.76e7 6.76e7 CTCF        coding:… nonsynonym… H     Y     CTCF_H->Y   
#> 2 chr17  7.58e6 7.58e6 TP53        intron:… nonsynonym… P     A     TP53_P->A   
#> 3 chr19  5.09e7 5.09e7 POLD1       coding   nonsynonym… T     M     POLD1_T->M  
#> # … with 1 more variable: is_driver <lgl>