Determines arm-level over-fragmentation patterns.

The fragmentation of a chromosome arm is assessed with a statistical test based on counting the size of the copy number segments mapping to the arm. This analysis works only at the level of clonal CNAs.

CNAqc counts, for every arm with lenght $L$ nucleotides:

- the number of mapped CNA segments shorter than a percentage of $L$; - the number of mapped CNA segments longer than a percentage of $L$.

A one-sided Binomial test is used to compute a p-value. In this way the test accounts for the difference in lenghts of the chromsome arms; a p-value per arm is reported and adjusted for multiple hyoptheses (Bonferroni).

detect_arm_overfragmentation(
  x,
  alpha = 0.01,
  genome_percentage_cutoff = 0.2,
  minimum_segments_for_testing = 10
)

Arguments

x

A CNAqc object.

alpha

Confidence level for the tests, for instance 0.05.

genome_percentage_cutoff

Segments are considered long or short depending on whether they are longer (in basepairs) than genome_percentage_cutoff * L bases, where L is the arm length for the reference genome. Default is 0.2 (twenty percent).

minimum_segments_for_testing

Smallest number of segments required to actually test a certain arm Default is 10 segments. This number influences the correction for mulitple hypothesis testing.

Value

A CNAqc objectwith the results.

Examples

data('example_dataset_CNAqc')
x = init(mutations = example_dataset_CNAqc$mutations, cna = example_dataset_CNAqc$cna, purity = example_dataset_CNAqc$purity)
#> 
#> ── CNAqc - CNA Quality Check ───────────────────────────────────────────────────
#> 
#> ℹ Using reference genome coordinates for: GRCh38.
#> ℹ Drivers are annotated, but 'gene' column is missing, using mutation location.
#> ✔ Fortified calls for 12963 somatic mutations: 12963 SNVs (100%) and 0 indels.
#> ! CNAs have no CCF, assuming clonal CNAs (CCF = 1).
#> ✔ Fortified CNAs for 267 segments: 267 clonal and 0 subclonal.
#> ✔ 12963 mutations mapped to clonal CNAs.

x = detect_arm_overfragmentation(x)
#> ℹ One-tailed Binomial test: 8 tests, alpha 0.01. Short segments: 0.2% of the reference arm.
#> ✔ chr7p,  p = 1.71798691840001e-24 ~ 34 segments, 34 short.
#> ✔ chr1p,  p = 1.62738995200002e-15 ~ 24 segments, 23 short.
#> ✔ chr11q,  p = 2.4788992e-07 ~ 14 segments, 12 short.
#> ✔ chr1q,  p = 4.34176000000001e-08 ~ 13 segments, 12 short.
#> ✔ chr12q,  p = 4.34176000000001e-08 ~ 13 segments, 12 short.
#> ✔ chr3q,  p = 4.52608e-06 ~ 12 segments, 10 short.
#> ✔ chr7q,  p = 4.52608e-06 ~ 12 segments, 10 short.
#> ✔ chr8p,  p = 9.21599999999998e-07 ~ 11 segments, 10 short.
#> ℹ 8 significantly overfragmented chromosome arms (alpha level 0.01).

# Report to console
print(x)
#> ── [ CNAqc ]  12963 mutations in 267 segments (267 clonal, 0 subclonal). Genome 
#> 
#> ── Clonal CNAs 
#> 
#>    2:2  [n = 7478, L = 1483 Mb] ■■■■■■■■■■■■■■■■■■■■■■■■■■■  { CTCF }
#>    4:2  [n = 1893, L =  331 Mb] ■■■■■■■
#>    3:2  [n = 1625, L =  357 Mb] ■■■■■■
#>    2:1  [n = 1563, L =  420 Mb] ■■■■■■  { TTN }
#>    3:0  [n =  312, L =  137 Mb] ■
#>    2:0  [n =   81, L =   39 Mb]   { TP53 }
#>   16:2  [n =    4, L =    0 Mb] 
#>   25:2  [n =    2, L =    1 Mb] 
#>    3:1  [n =    2, L =    1 Mb] 
#>  106:1  [n =    1, L =    0 Mb] 
#> 
#> ℹ Sample Purity: 89% ~ Ploidy: 4.
#> 
#> ℹ There are 3 annotated driver(s) mapped to clonal CNAs.
#>          chr      from        to ref alt  DP NV       VAF driver_label is_driver
#>         chr2 179431633 179431634   C   T 117 77 0.6581197          TTN      TRUE
#>        chr16  67646006  67646007   C   T 120 54 0.4500000         CTCF      TRUE
#>        chr17   7577106   7577107   G   C  84 78 0.9285714         TP53      TRUE
#> ✔ Arm-level fragmentation analysis: 8 segments overfragmented.