Create a CNAqc object.

Creates a CNAqc object from a set of mutation and CNA calls, and a tumour purity value. The resulting object contains the mutations mapped into CNA segments, and allows for the computation of the QC metrics available in the CNAqc package. CNAqc supports `hg19` coordinates.

init(snvs, cna, purity, ref = "GRCh38")

Arguments

snvs
A data.frame of mutations with the following fields available: `chr`, `from`, `to` as `hg19` chromosome coordinates. `ref` and `alt` for the reference and alternative alleles, and `DP`, `NV` and `VAF` for the depth (total number of reads with both reference and alternative), the number of variants with the alternative allele and the allele frequency (VAF = NV/DP). Chromosome names must be in the format `chr1`, `chr2`, etc.; alleles should be characters and all other fields numeric.

cna
A data.frame of CNA with the following fields available: `chr`, `from`, `to` as `hg19` chromosome coordinates. `Major` and `minor` for the number of copies of the major allele, and the minor (B-allele). A `CCF` column can be used as a real-value in between 0 and 1 to represent Cancer Cell Fractions; if this is not available `CCF = 1` is set and all calls will refer to clonal segments. Otherwise, segments with `CCF<1` would be considered subclonal CNAs.

purity
Value in between `0` and `1` to represent the proportion of actual tumour content (sometimes called "cellularity").

ref
The reference genome (either "hg19" or "GRCh38"); the default is "GRCh38".

Value

A CNAqc object of class `cnaqc`, with S3 methods for printing, plotting and analyzing data.

Examples

data('example_dataset_CNAqc', package = 'CNAqc')
print(example_dataset_CNAqc)

#> $snvs
#> # A tibble: 12,963 x 13
#>    chr     from     to ref   alt   FILTER    DP    NV    VAF ANNOVAR_FUNCTION
#>    <chr>  <dbl>  <dbl> <chr> <chr> <chr>  <dbl> <dbl>  <dbl> <chr>           
#>  1 chr1  1.03e6 1.03e6 T     G     PASS      60     6 0.1    UTR5            
#>  2 chr1  2.25e6 2.25e6 A     C     PASS     127     9 0.0709 intergenic      
#>  3 chr1  2.46e6 2.46e6 G     A     PASS     156    65 0.417  upstream        
#>  4 chr1  2.73e6 2.73e6 T     C     PASS     180    90 0.5    downstream      
#>  5 chr1  2.76e6 2.76e6 C     T     PASS     183    61 0.333  intergenic      
#>  6 chr1  2.77e6 2.77e6 C     T     PASS     203   130 0.640  intergenic      
#>  7 chr1  2.94e6 2.94e6 C     T     PASS     228   132 0.579  intergenic      
#>  8 chr1  2.98e6 2.98e6 C     T     PASS     196    85 0.434  ncRNA_exonic    
#>  9 chr1  3.39e6 3.39e6 T     G     PASS     124     6 0.0484 intronic        
#> 10 chr1  3.50e6 3.50e6 G     A     PASS      88    10 0.114  intronic        
#> # … with 12,953 more rows, and 3 more variables: GENE <chr>, is_driver <lgl>,
#> #   driver_label <chr>
#> 
#> $cna
#> # A tibble: 267 x 7
#>    chr       from       to  length covRatio Major minor
#>    <chr>    <int>    <int>   <int>    <dbl> <dbl> <dbl>
#>  1 chr1    840009  1689987  849979     1.19     3     2
#>  2 chr1   1689988  1815015  125028     1.26     3     2
#>  3 chr1   1815016  9799969 7984954     1.19     3     2
#>  4 chr1  10479910 12079917 1600008     1.19     3     2
#>  5 chr1  12079917 12154980   75064     1.24     3     2
#>  6 chr1  12154981 12839977  684997     1.19     3     2
#>  7 chr1  13780016 17790026 4010011     1.19     3     2
#>  8 chr1  17849962 21080067 3230106     1.19     3     2
#>  9 chr1  21080068 21559998  479931     1.26     3     2
#> 10 chr1  21559998 24830001 3270004     1.19     3     2
#> # … with 257 more rows
#> 
#> $purity
#> [1] 0.89
#> 
#> $reference
#> [1] "hg19"
#> 

x = init(snvs = example_dataset_CNAqc$snvs, cna = example_dataset_CNAqc$cna, purity = example_dataset_CNAqc$purity)

#>  [ CNAqc - CNA Quality Check ] 
#> 

#> ℹ Using reference genome coordinates for: GRCh38.

#> ℹ Drivers are annotated, but 'gene' column is missing, using mutation location.

#> ! Missing CCF column from CNA calls, adding CCF = 1 assuming clonal CNA calls.

#> ℹ Input n = 12963 mutations for 267 CNA segments (267 clonal, 0 subclonal)

#> 

#> ✓ Mapped n = 12963 mutations to clonal segments (100% of input)

print(x)

#> ── [ CNAqc ]  n = 12963 mutations in 267 segments (267 clonal + 0 subclonal). Ge

#> 
#>    2:2  [n = 7478, L = 1483 Mb] ■■■■■■■■■■■■■■■■■■■■■■■■■■■  { CTCF }
#>    4:2  [n = 1893, L =  331 Mb] ■■■■■■■
#>    3:2  [n = 1625, L =  357 Mb] ■■■■■■
#>    2:1  [n = 1563, L =  420 Mb] ■■■■■■  { TTN }
#>    3:0  [n =  312, L =  137 Mb] ■
#>    2:0  [n =   81, L =   39 Mb]   { TP53 }
#>   16:2  [n =    4, L =    0 Mb] 
#>   25:2  [n =    2, L =    1 Mb] 
#>    3:1  [n =    2, L =    1 Mb] 
#>  106:1  [n =    1, L =    0 Mb] 
#>   26:2  [n =    1, L =    0 Mb] 
#>   99:1  [n =    1, L =    0 Mb] 
#> 

#> ℹ Sample Purity: 89% ~ Ploidy: 4.

#>