Subset calls by segments' karyotype.

For an object already created, it subsets the calls to those that involve segments with a certain karyotype.

subset_by_segment_karyotype(x, karyotypes)

Arguments

x
An object of class cnaqc, created by the init function.

karyotype
A list of karyotype ids in "Major:minor" notation (e.g., "1:1", "2,1", ...) that will be retained.

Value

An object of class cnaqc, created by the init function.

Examples

data('example_dataset_CNAqc', package = 'CNAqc')
x = init(example_dataset_CNAqc$snvs, example_dataset_CNAqc$cna,example_dataset_CNAqc$purity)

#>  [ CNAqc - CNA Quality Check ] 
#> 

#> ℹ Using reference genome coordinates for: GRCh38.

#> ℹ Drivers are annotated, but 'gene' column is missing, using mutation location.

#> ! Missing CCF column from CNA calls, adding CCF = 1 assuming clonal CNA calls.

#> ℹ Input n = 12963 mutations for 267 CNA segments (267 clonal, 0 subclonal)

#> 

#> ✔ Mapped n = 12963 mutations to clonal segments (100% of input)

subset_by_segment_karyotype(x, '2:2')

#>  [ CNAqc - CNA Quality Check ] 
#> 

#> ℹ Using reference genome coordinates for: GRCh38.

#> ! Missing CCF column from CNA calls, adding CCF = 1 assuming clonal CNA calls.

#> ℹ Input n = 12963 mutations for 80 CNA segments (80 clonal, 0 subclonal)

#> Warning: [CNAqc] a karyotype column is present in CNA calls, and will be overwritten

#> ✖ n = 5485 mutations  cannot be mapped to segments and will be removed.

#> 

#> ✔ Mapped n = 7478 mutations to clonal segments (100% of input)

#> ── [ CNAqc ]  n = 7478 mutations in 80 segments (80 clonal + 0 subclonal). Genom

#> 
#>  2:2  [n = 7478, L = 1483 Mb] ■■■■■■■■■■■■■■■■■■■■■■■■■■■  { CTCF }
#> 

#> ℹ Sample Purity: 89% ~ Ploidy: 4.

#> 

#> ℹ There are 1 annotated driver(s).

#>          chr     from       to ref alt  DP NV  VAF driver_label is_driver
#>        chr16 67646006 67646007   C   T 120 54 0.45         CTCF      TRUE

subset_by_segment_karyotype(x, '2:1')

#>  [ CNAqc - CNA Quality Check ] 
#> 

#> ℹ Using reference genome coordinates for: GRCh38.

#> ! Missing CCF column from CNA calls, adding CCF = 1 assuming clonal CNA calls.

#> ℹ Input n = 12963 mutations for 49 CNA segments (49 clonal, 0 subclonal)

#> Warning: [CNAqc] a karyotype column is present in CNA calls, and will be overwritten

#> ✖ n = 11400 mutations  cannot be mapped to segments and will be removed.

#> 

#> ✔ Mapped n = 1563 mutations to clonal segments (100% of input)

#> ── [ CNAqc ]  n = 1563 mutations in 49 segments (49 clonal + 0 subclonal). Genom

#> 
#>  2:1  [n = 1563, L =  420 Mb] ■■■■■■■■■■■■■■■■■■■■■■■■■■■  { TTN }
#> 

#> ℹ Sample Purity: 89% ~ Ploidy: 3.

#> 

#> ℹ There are 1 annotated driver(s).

#>         chr      from        to ref alt  DP NV       VAF driver_label is_driver
#>        chr2 179431633 179431634   C   T 117 77 0.6581197          TTN      TRUE